refactor(server): split home_page.rs into submodules
home_page.rs was ~6,258 lines. Split into focused files under
home_page/ following Edition 2024 style (root file + sibling
directory, no mod.rs):

- assets.rs   — CSS, JS, and PEER_CSS string constants
- cards.rs    — dashboard card builders
- estimator.rs — regression/reliability chart builders and RegKind enum
- favicon.rs  — SVG favicon renderer
- peer_detail.rs — per-peer detail page HTML builder

The root home_page.rs retains RABBIT_SVG_PATH (accessed by the
sibling path_handlers.rs via super::home_page::), the two handler
fns (homepage, peer_detail), and the full test module with explicit
submodule imports. Source-scrape pin tests updated to point at the
new submodule paths. No behavior changes.
