+++ crates/core/src/server/home_page.rs
@@ -3,12 +3,95 @@
 //! Renders a card-based dashboard showing connection status, peers, subscriptions,
 //! and operation stats. Styled to match the global telemetry dashboard.
 
-use std::fmt::Write;
+mod assets;
+mod cards;
+mod estimator;
+mod favicon;
+mod peer_detail;
 
 use axum::extract::Path;
 use axum::response::{Html, IntoResponse};
 
-use crate::node::network_status::{self, format_ago, format_duration, html_escape};
+// Re-exported so child submodules can reach them via `use super::*;`.
+// Safe to re-export at pub(super) because they are pub in their origin.
+pub(super) use crate::node::network_status::{self, format_ago, format_duration, html_escape};
+pub(super) use std::fmt::Write;
+
+use assets::{CSS, JS};
+use cards::{
+    build_ban_list_card, build_contracts_card, build_governance_card, build_ops_card,
+    build_peers_card, build_status_card, build_transfer_card,
+};
+use favicon::build_favicon_data_uri;
+use peer_detail::peer_detail_html;
+
+/// Freenet rabbit silhouette SVG path, derived from freenet_logo.svg.
+/// Used for the favicon with a solid color fill (no gradient) so the
+/// connection status color is immediately visible at favicon size.
+/// Kept in root (not favicon.rs) so `path_handlers.rs` can reach it as
+/// `super::home_page::RABBIT_SVG_PATH` without an extra re-export.
+pub(super) const RABBIT_SVG_PATH: &str = concat!(
+    "M358.864 40.470C358.605 40.728 354.143 42.467 348.947 44.334",
+    "C284.621 67.446 232.573 113.729 201.443 175.500",
+    "C193.895 190.478 184.375 213.708 185.375 214.708",
+    "C185.621 214.954 187.715 211.857 190.030 207.827",
+    "C211.190 170.984 229.863 146.093 255.968 119.933",
+    "C274.854 101.008 282.998 94.207 302.034 81.466",
+    "C334.671 59.621 367.531 47.376 401.250 44.492",
+    "L409.000 43.829 408.984 47.165",
+    "C408.958 52.704 405.255 68.515 401.010 81.213",
+    "C382.392 136.898 338.799 184.709 277.000 217.224",
+    "C271.225 220.263 263.913 223.788 260.750 225.058",
+    "C254.629 227.517 254.126 228.307 256.511 231.712",
+    "C258.282 234.241 258.484 234.089 249.500 237.002",
+    "C226.868 244.341 200.420 256.771 183.918 267.825",
+    "C173.918 274.522 156.961 289.225 158.000 290.296",
+    "C158.275 290.579 163.450 287.694 169.500 283.883",
+    "C175.550 280.073 186.125 274.083 193.000 270.573",
+    "C264.905 233.856 345.414 226.155 422.387 248.633",
+    "C434.634 252.210 468.194 264.823 465.830 264.961",
+    "C465.461 264.982 459.741 263.440 453.118 261.534",
+    "C422.666 252.769 376.068 246.967 347.500 248.384",
+    "C320.590 249.719 284.052 255.527 283.798 258.510",
+    "C283.694 259.727 291.796 264.541 298.477 267.231",
+    "C306.163 270.326 319.360 270.612 338.812 268.103",
+    "C385.602 262.070 433.627 269.250 469.963 287.712",
+    "C475.721 290.638 480.874 293.474 481.416 294.016",
+    "C482.061 294.661 476.225 295.004 464.450 295.010",
+    "C407.520 295.043 349.853 308.084 300.500 332.086",
+    "C290.412 336.992 272.833 346.834 271.147 348.519",
+    "C269.278 350.389 273.301 349.263 283.966 344.933",
+    "C302.548 337.389 332.479 327.629 351.000 323.074",
+    "C386.266 314.400 413.893 311.393 450.000 312.298",
+    "C474.559 312.914 491.602 315.535 509.306 321.421",
+    "C520.784 325.236 519.954 325.640 504.924 323.552",
+    "C419.615 311.701 330.506 332.225 238.000 385.033",
+    "C224.991 392.460 221.855 394.386 200.762 407.913",
+    "C184.591 418.282 178.817 420.978 172.750 420.990",
+    "C167.060 421.002 164.441 418.869 163.413 413.387",
+    "C160.912 400.055 178.394 366.762 202.136 339.644",
+    "L206.387 334.788 197.443 335.644",
+    "C183.073 337.019 158.519 336.519 150.152 334.681",
+    "C132.833 330.876 120.785 321.947 117.439 310.437",
+    "C112.326 292.850 123.492 270.717 146.912 252.015",
+    "C154.528 245.934 155.702 244.562 156.798 240.464",
+    "C158.983 232.296 168.599 206.900 174.208 194.482",
+    "C184.044 172.710 197.989 150.083 213.332 131.000",
+    "C229.597 110.770 255.612 87.415 277.277 73.590",
+    "C286.990 67.393 310.855 55.436 323.062 50.653",
+    "C335.623 45.730 360.750 38.583 358.864 40.470",
+    "M375.000 209.596",
+    "C430.712 216.655 477.541 237.609 509.241 269.661",
+    "C514.049 274.523 518.765 279.625 519.721 281.000",
+    "C521.404 283.419 521.347 283.408 517.980 280.669",
+    "C500.484 266.434 486.556 257.516 468.000 248.668",
+    "C452.323 241.193 438.766 236.261 424.000 232.663",
+    "C395.297 225.667 374.955 223.024 349.705 223.010",
+    "C333.212 223.000 332.865 222.956 330.455 220.545",
+    "C329.105 219.195 328.000 217.046 328.000 215.768",
+    "C328.000 212.845 330.558 209.125 333.357 207.978",
+    "C336.189 206.817 360.536 207.763 375.000 209.596",
+);
 
 /// Handler for `GET /` — returns a self-contained HTML dashboard.
 pub(super) async fn homepage() -> impl IntoResponse {
@@ -168,4124 +251,23 @@ fn homepage_html() -> String {
     )
 }
 
-/// Freenet rabbit silhouette SVG path, derived from freenet_logo.svg.
-/// Used for the favicon with a solid color fill (no gradient) so the
-/// connection status color is immediately visible at favicon size.
-pub(super) const RABBIT_SVG_PATH: &str = concat!(
-    "M358.864 40.470C358.605 40.728 354.143 42.467 348.947 44.334",
-    "C284.621 67.446 232.573 113.729 201.443 175.500",
-    "C193.895 190.478 184.375 213.708 185.375 214.708",
-    "C185.621 214.954 187.715 211.857 190.030 207.827",
-    "C211.190 170.984 229.863 146.093 255.968 119.933",
-    "C274.854 101.008 282.998 94.207 302.034 81.466",
-    "C334.671 59.621 367.531 47.376 401.250 44.492",
-    "L409.000 43.829 408.984 47.165",
-    "C408.958 52.704 405.255 68.515 401.010 81.213",
-    "C382.392 136.898 338.799 184.709 277.000 217.224",
-    "C271.225 220.263 263.913 223.788 260.750 225.058",
-    "C254.629 227.517 254.126 228.307 256.511 231.712",
-    "C258.282 234.241 258.484 234.089 249.500 237.002",
-    "C226.868 244.341 200.420 256.771 183.918 267.825",
-    "C173.918 274.522 156.961 289.225 158.000 290.296",
-    "C158.275 290.579 163.450 287.694 169.500 283.883",
-    "C175.550 280.073 186.125 274.083 193.000 270.573",
-    "C264.905 233.856 345.414 226.155 422.387 248.633",
-    "C434.634 252.210 468.194 264.823 465.830 264.961",
-    "C465.461 264.982 459.741 263.440 453.118 261.534",
-    "C422.666 252.769 376.068 246.967 347.500 248.384",
-    "C320.590 249.719 284.052 255.527 283.798 258.510",
-    "C283.694 259.727 291.796 264.541 298.477 267.231",
-    "C306.163 270.326 319.360 270.612 338.812 268.103",
-    "C385.602 262.070 433.627 269.250 469.963 287.712",
-    "C475.721 290.638 480.874 293.474 481.416 294.016",
-    "C482.061 294.661 476.225 295.004 464.450 295.010",
-    "C407.520 295.043 349.853 308.084 300.500 332.086",
-    "C290.412 336.992 272.833 346.834 271.147 348.519",
-    "C269.278 350.389 273.301 349.263 283.966 344.933",
-    "C302.548 337.389 332.479 327.629 351.000 323.074",
-    "C386.266 314.400 413.893 311.393 450.000 312.298",
-    "C474.559 312.914 491.602 315.535 509.306 321.421",
-    "C520.784 325.236 519.954 325.640 504.924 323.552",
-    "C419.615 311.701 330.506 332.225 238.000 385.033",
-    "C224.991 392.460 221.855 394.386 200.762 407.913",
-    "C184.591 418.282 178.817 420.978 172.750 420.990",
-    "C167.060 421.002 164.441 418.869 163.413 413.387",
-    "C160.912 400.055 178.394 366.762 202.136 339.644",
-    "L206.387 334.788 197.443 335.644",
-    "C183.073 337.019 158.519 336.519 150.152 334.681",
-    "C132.833 330.876 120.785 321.947 117.439 310.437",
-    "C112.326 292.850 123.492 270.717 146.912 252.015",
-    "C154.528 245.934 155.702 244.562 156.798 240.464",
-    "C158.983 232.296 168.599 206.900 174.208 194.482",
-    "C184.044 172.710 197.989 150.083 213.332 131.000",
-    "C229.597 110.770 255.612 87.415 277.277 73.590",
-    "C286.990 67.393 310.855 55.436 323.062 50.653",
-    "C335.623 45.730 360.750 38.583 358.864 40.470",
-    "M375.000 209.596",
-    "C430.712 216.655 477.541 237.609 509.241 269.661",
-    "C514.049 274.523 518.765 279.625 519.721 281.000",
-    "C521.404 283.419 521.347 283.408 517.980 280.669",
-    "C500.484 266.434 486.556 257.516 468.000 248.668",
-    "C452.323 241.193 438.766 236.261 424.000 232.663",
-    "C395.297 225.667 374.955 223.024 349.705 223.010",
-    "C333.212 223.000 332.865 222.956 330.455 220.545",
-    "C329.105 219.195 328.000 217.046 328.000 215.768",
-    "C328.000 212.845 330.558 209.125 333.357 207.978",
-    "C336.189 206.817 360.536 207.763 375.000 209.596",
-);
-
-/// Build a `data:` URI for an SVG favicon colored by connection status.
-///
-/// Colors follow issue #3287 (match order = priority):
-/// 1. Grey: starting up (no snapshot yet)
-/// 2. Blue: connected (any open connections — healthy state wins)
-/// 3. Dark red: NAT traversal failing (all attempts failed)
-/// 4. Red: connection failures present
-/// 5. Amber: attempting to connect (fallback)
-fn build_favicon_data_uri(snap: &Option<network_status::NetworkStatusSnapshot>) -> String {
-    // Color is pre-encoded for data URI (# → %23) to avoid scanning the entire SVG.
-    let color = match snap {
-        None => "%239e9e9e",                              // grey — starting up
-        Some(s) if s.open_connections > 0 => "%230abab5", // teal — connected
-        Some(s) if s.nat_stats.attempts > 0 && s.nat_stats.successes == 0 => "%238b0000", // dark red — NAT problems
-        Some(s) if !s.failures.is_empty() => "%23f44336", // red — connection issues
-        Some(_) => "%23fbbf24",                           // amber — connecting
-    };
-
-    format!(
-        "data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 640 471'>\
-         <path d='{path}' fill='{color}' fill-rule='evenodd'/></svg>",
-        path = RABBIT_SVG_PATH,
-        color = color,
-    )
-}
-
-fn build_status_card(snap: &Option<network_status::NetworkStatusSnapshot>) -> String {
-    let Some(snap) = snap else {
-        return r#"<div class="card">
-            <h2>Connection Status</h2>
-            <div class="status-row"><span class="dot dot-yellow"></span> Starting up...</div>
-            <div class="spinner"></div>
-        </div>"#
-            .to_string();
-    };
-
-    // Health banner — the primary "everything looks good" indicator
-    let health_banner = match snap.health {
-        network_status::HealthLevel::Healthy => {
-            let n = snap.open_connections;
-            let label = if n == 1 { "peer" } else { "peers" };
-            format!(
-                r#"<div class="health-banner health-good">
-                    <span class="health-icon">&#x2714;</span>
-                    <span>Node is healthy — connected to {n} {label}</span>
-                </div>"#,
-            )
-        }
-        network_status::HealthLevel::Degraded => {
-            let detail = if snap.gateway_only {
-                "Only connected to gateways — no peer-to-peer connections yet"
-            } else {
-                "Connected but NAT traversal is failing"
-            };
-            format!(
-                r#"<div class="health-banner health-degraded">
-                    <span class="health-icon">&#x26A0;</span>
-                    <span>{detail}</span>
-                </div>"#,
-            )
-        }
-        network_status::HealthLevel::Connecting => r#"<div class="health-banner health-connecting">
-                <span class="health-icon">&#x231B;</span>
-                <span>Connecting to the network...</span>
-            </div>"#
-            .to_string(),
-        network_status::HealthLevel::Trouble => {
-            let has_version_mismatch = snap
-                .failures
-                .iter()
-                .any(|f| f.reason_html.contains("Version mismatch"));
-            let detail = if has_version_mismatch {
-                "Version mismatch — update required"
-            } else {
-                "Unable to connect — check firewall and network settings"
-            };
-            format!(
-                r#"<div class="health-banner health-trouble">
-                    <span class="health-icon">&#x2716;</span>
-                    <span>{detail}</span>
-                </div>"#,
-            )
-        }
-    };
-
-    // External address info (shown once discovered via NAT traversal)
-    let external_addr_html = if let Some(addr) = snap.external_address {
-        format!(
-            r#"<p class="external-addr">External address: <code>{ip}</code> &mdash; UDP port: <code>{port}</code></p>"#,
-            ip = addr.ip(),
-            port = addr.port(),
-        )
-    } else if snap.open_connections > 0 {
-        r#"<p class="external-addr muted">External address: discovering...</p>"#.to_string()
-    } else {
-        String::new()
-    };
-
-    // Ring stats row: connection count, hosted contracts, connection attempts
-    let ring_stats_html = format!(
-        r#"<div class="metrics-row">
-            <div class="metric-tile">
-                <span class="metric-value">{conns}</span>
-                <span class="metric-label">Ring peers</span>
-            </div>
-            <div class="metric-tile">
-                <span class="metric-value">{hosted}</span>
-                <span class="metric-label">Hosted contracts</span>
-            </div>
-            <div class="metric-tile">
-                <span class="metric-value">{attempts}</span>
-                <span class="metric-label">Conn attempts</span>
-            </div>
-        </div>"#,
-        conns = snap.ring_stats.connection_count,
-        hosted = snap.ring_stats.hosted_contracts,
-        attempts = snap.connection_attempts,
-    );
-
-    // UPDATE rate-limiter stats: only shown once the limiter has seen
-    // traffic, so idle nodes stay uncluttered. A non-zero "Rate-limited"
-    // or "Capacity-dropped" count is the operator's signal that the
-    // per-(sender, contract) UPDATE limiter is dropping relayed traffic.
-    let rate_limit_html = if snap.ring_stats.updates_accepted > 0
-        || snap.ring_stats.updates_rate_limited > 0
-        || snap.ring_stats.updates_capacity_dropped > 0
-    {
-        format!(
-            r#"<div class="metrics-row">
-            <div class="metric-tile">
-                <span class="metric-value">{accepted}</span>
-                <span class="metric-label">UPDATEs relayed</span>
-            </div>
-            <div class="metric-tile">
-                <span class="metric-value">{rate_limited}</span>
-                <span class="metric-label">Rate-limited</span>
-            </div>
-            <div class="metric-tile">
-                <span class="metric-value">{capacity_dropped}</span>
-                <span class="metric-label">Capacity-dropped</span>
-            </div>
-        </div>"#,
-            accepted = snap.ring_stats.updates_accepted,
-            rate_limited = snap.ring_stats.updates_rate_limited,
-            capacity_dropped = snap.ring_stats.updates_capacity_dropped,
-        )
-    } else {
-        String::new()
-    };
-
-    let spinner = if snap.open_connections == 0 {
-        r#"<div class="spinner"></div>"#
-    } else {
-        ""
-    };
-
-    // Gateway-only warning (only when not connected to any peers)
-    let gateway_warning = if snap.gateway_only {
-        format!(
-            r#"<div class="warning">
-                <strong>Firewall likely blocking incoming connections</strong> on UDP port <code>{port}</code>.
-                <ul>
-                    <li>Configure your router to forward UDP port <code>{port}</code> to this computer.</li>
-                    <li>Check that no software firewall (ufw, iptables, Windows Defender) is blocking Freenet.</li>
-                </ul>
-            </div>"#,
-            port = snap.listening_port
-        )
-    } else {
-        String::new()
-    };
-
-    // NAT stats with rolling trend
-    let nat_html = if snap.nat_stats.attempts > 0 {
-        let all_failed = snap.nat_stats.successes == 0;
-        let class = if all_failed { " nat-fail" } else { "" };
-        let extra = if all_failed && !snap.gateway_only {
-            format!(
-                r#"<p class="nat-advice">All NAT traversal attempts have failed. Try forwarding UDP port <code>{}</code> on your router.</p>"#,
-                snap.listening_port
-            )
-        } else {
-            String::new()
-        };
-
-        // Rolling trend: recent window stats; only show verdict when truly blocked
-        let (recent, verdict) = if snap.nat_stats.recent_attempts > 0 {
-            let rs = snap.nat_stats.recent_successes;
-            let ra = snap.nat_stats.recent_attempts;
-            let verdict = if rs == 0 && snap.nat_stats.successes == 0 {
-                r#" <span class="nat-verdict nat-verdict-bad">Port may be blocked</span>"#
-                    .to_string()
-            } else {
-                String::new()
-            };
-            (
-                format!(r#" <span class="nat-recent">({rs}/{ra} recent)</span>"#),
-                verdict,
-            )
-        } else {
-            (String::new(), String::new())
-        };
-
-        format!(
-            r#"<p class="nat-stat{class}">NAT hole punching: {s}/{a} successful{recent} {verdict}</p>{extra}"#,
-            class = class,
-            s = snap.nat_stats.successes,
-            a = snap.nat_stats.attempts,
-            recent = recent,
-            verdict = verdict,
-            extra = extra,
-        )
-    } else if snap.open_connections == 0 {
-        r#"<p class="nat-stat">No NAT traversal attempts yet</p>"#.to_string()
-    } else {
-        String::new()
-    };
-
-    // Failure diagnostics — demoted when connected (muted style, collapsed)
-    let failures_html = if !snap.failures.is_empty() {
-        let mut items = String::new();
-        for f in &snap.failures {
-            items.push_str(&format!(
-                "<li><code>{}</code>: {}</li>",
-                f.address, f.reason_html
-            ));
-        }
-        if snap.open_connections > 0 {
-            // Demoted: muted style when node is otherwise connected
-            format!(
-                r#"<details class="diagnostics-muted">
-                    <summary>{n} recent connection attempt(s) failed <span class="muted-hint">(normal)</span></summary>
-                    <ul>{items}</ul>
-                </details>"#,
-                n = snap.failures.len(),
-                items = items,
-            )
-        } else {
-            // Prominent: when not connected, failures are actionable
-            format!(
-                r#"<div class="diagnostics">
-                    <h3>Connection Issues</h3>
-                    <ul>{items}</ul>
-                    <p class="attempts">Attempted {attempts} connection(s) over {elapsed}. Retrying...</p>
-                </div>"#,
-                items = items,
-                attempts = snap.connection_attempts,
-                elapsed = format_duration(snap.elapsed_secs),
-            )
-        }
-    } else if snap.open_connections == 0 && snap.connection_attempts > 0 {
-        format!(
-            r#"<p class="attempts">Attempted {} connection(s) over {}. Retrying...</p>"#,
-            snap.connection_attempts,
-            format_duration(snap.elapsed_secs),
-        )
-    } else {
-        String::new()
-    };
-
-    format!(
-        r#"<div class="card">
-            <h2>Connection Status</h2>
-            {health_banner}
-            {ring_stats_html}
-            {rate_limit_html}
-            {external_addr_html}
-            {spinner}
-            {gateway_warning}
-            {nat_html}
-            {failures_html}
-        </div>"#,
-        health_banner = health_banner,
-        ring_stats_html = ring_stats_html,
-        rate_limit_html = rate_limit_html,
-        external_addr_html = external_addr_html,
-        spinner = spinner,
-        gateway_warning = gateway_warning,
-        nat_html = nat_html,
-        failures_html = failures_html,
-    )
-}
-
-/// Format bytes as a human-readable string (e.g., "1.2 MB").
-fn format_bytes(bytes: u64) -> String {
-    const KB: u64 = 1024;
-    const MB: u64 = 1024 * KB;
-    const GB: u64 = 1024 * MB;
-
-    if bytes >= GB {
-        format!("{:.1} GB", bytes as f64 / GB as f64)
-    } else if bytes >= MB {
-        format!("{:.1} MB", bytes as f64 / MB as f64)
-    } else if bytes >= KB {
-        format!("{:.1} KB", bytes as f64 / KB as f64)
-    } else {
-        format!("{bytes} B")
-    }
-}
-
-fn build_transfer_card(snap: &Option<network_status::NetworkStatusSnapshot>) -> String {
-    let Some(snap) = snap else {
-        return String::new();
-    };
-    // Always show the card once the node has been up for more than a few
-    // seconds — initial connection flapping shouldn't make the panel
-    // appear/disappear rhythmically during auto-refresh.
-    if snap.bytes_uploaded == 0 && snap.bytes_downloaded == 0 && snap.elapsed_secs < 10 {
-        return String::new();
-    }
-
-    let ts = &snap.transport_snapshot;
-
-    let peak_tput = if ts.peak_throughput_bps > 0 {
-        format!(
-            " <span class=\"transfer-detail\">(peak {}/s)</span>",
-            format_bytes(ts.peak_throughput_bps)
-        )
-    } else {
-        String::new()
-    };
-
-    let rtt_str = if ts.avg_rtt_us > 0 {
-        format!(
-            r#"<div class="transfer-stat">
-                <span class="transfer-label">RTT (avg/min/max)</span>
-                <span class="transfer-value">{avg}ms / {min}ms / {max}ms</span>
-            </div>"#,
-            avg = format_args!("{:.1}", ts.avg_rtt_us as f64 / 1000.0),
-            min = format_args!("{:.1}", ts.min_rtt_us as f64 / 1000.0),
-            max = format_args!("{:.1}", ts.max_rtt_us as f64 / 1000.0),
-        )
-    } else {
-        String::new()
-    };
-
-    let cwnd_str = if ts.avg_cwnd_bytes > 0 {
-        format!(
-            r#"<div class="transfer-stat">
-                <span class="transfer-label">cwnd (avg/peak/min)</span>
-                <span class="transfer-value">{avg} / {peak} / {min}</span>
-            </div>"#,
-            avg = format_bytes(ts.avg_cwnd_bytes as u64),
-            peak = format_bytes(ts.peak_cwnd_bytes as u64),
-            min = format_bytes(ts.min_cwnd_bytes as u64),
-        )
-    } else {
-        String::new()
-    };
-
-    let slowdown_str = if ts.slowdowns_triggered > 0 {
-        format!(
-            r#"<div class="transfer-stat">
-                <span class="transfer-label">LEDBAT slowdowns</span>
-                <span class="transfer-value">{s}</span>
-            </div>"#,
-            s = ts.slowdowns_triggered,
-        )
-    } else {
-        String::new()
-    };
-
-    let xfer_str = if ts.transfers_completed > 0 || ts.transfers_failed > 0 {
-        format!(
-            r#"<div class="transfer-stat">
-                <span class="transfer-label">Transfers (avg time)</span>
-                <span class="transfer-value">{ok} ok / {fail} fail <span class="transfer-detail">({avg}s avg)</span></span>
-            </div>"#,
-            ok = ts.transfers_completed,
-            fail = ts.transfers_failed,
-            avg = format_args!("{:.3}", ts.avg_transfer_time_ms as f64 / 1000.0),
-        )
-    } else {
-        String::new()
-    };
-
-    format!(
-        r#"<div class="card">
-            <h2>Data Transfer</h2>
-            <div class="transfer-stat">
-                <span class="transfer-label">Uploaded</span>
-                <span class="transfer-value">{uploaded}{peak_tput}</span>
-            </div>
-            <div class="transfer-stat">
-                <span class="transfer-label">Downloaded</span>
-                <span class="transfer-value">{downloaded}</span>
-            </div>
-            {xfer_str}
-            {rtt_str}
-            {cwnd_str}
-            {slowdown_str}
-        </div>"#,
-        uploaded = format_bytes(snap.bytes_uploaded),
-        peak_tput = peak_tput,
-        downloaded = format_bytes(snap.bytes_downloaded),
-        xfer_str = xfer_str,
-        rtt_str = rtt_str,
-        cwnd_str = cwnd_str,
-        slowdown_str = slowdown_str,
-    )
-}
-
-fn build_peers_card(snap: &Option<network_status::NetworkStatusSnapshot>) -> String {
-    let Some(snap) = snap else {
-        return String::new();
-    };
-    if snap.peers.is_empty() && snap.open_connections == 0 {
-        return String::new();
-    }
-
-    let own_loc = snap
-        .own_location
-        .map(|l| format!(r#"<span class="own-loc">Your location: {:.4}</span>"#, l))
-        .unwrap_or_default();
-
-    if snap.peers.is_empty() {
-        return format!(
-            r#"<div class="card">
-                <div class="card-header"><h2>Network Peers</h2>{own_loc}</div>
-                <p class="empty">No peers connected</p>
-            </div>"#,
-            own_loc = own_loc,
-        );
-    }
-
-    let ring_svg = build_ring_svg(
-        snap.own_location,
-        &snap.peers,
-        Some(&snap.governance),
-        &snap.contracts,
-    );
-
-    let mut rows = String::new();
-    for p in &snap.peers {
-        let peer_type = if p.is_gateway { "Gateway" } else { "Peer" };
-        let loc_sort = p.location.map(|l| l.to_string()).unwrap_or_default();
-        let loc = p
-            .location
-            .map(|l| format!("{:.4}", l))
-            .unwrap_or_else(|| "—".to_string());
-        let sent = if p.bytes_sent > 0 {
-            format_bytes(p.bytes_sent)
-        } else {
-            "—".to_string()
-        };
-        let recv = if p.bytes_received > 0 {
-            format_bytes(p.bytes_received)
-        } else {
-            "—".to_string()
-        };
-        rows.push_str(&format!(
-            r#"<tr class="peer-row" onclick="window.location='/peer/{addr_enc}'"><td data-sort="{addr_enc}"><code>{addr}</code></td><td data-sort="{loc_sort}">{loc}</td><td data-sort="{ptype}">{ptype}</td><td data-sort="{bytes_sent}">{sent}</td><td data-sort="{bytes_recv}">{recv}</td><td data-sort="{conn_secs}">{connected}</td></tr>"#,
-            addr_enc = html_escape(&p.address.to_string()),
-            addr = p.address,
-            loc_sort = loc_sort,
-            loc = loc,
-            ptype = peer_type,
-            bytes_sent = p.bytes_sent,
-            sent = sent,
-            bytes_recv = p.bytes_received,
-            recv = recv,
-            conn_secs = p.connected_secs,
-            connected = format_duration(p.connected_secs),
-        ));
-    }
-
-    format!(
-        r#"<div class="card">
-            <div class="card-header"><h2>Network Peers</h2>{own_loc}</div>
-            {ring_svg}
-            <div class="table-wrap">
-                <table class="sortable" data-table-id="peers">
-                    <thead><tr><th data-sort-type="text">Address</th><th data-sort-type="num">Location</th><th data-sort-type="text">Type</th><th data-sort-type="num">Sent</th><th data-sort-type="num">Recv</th><th data-sort-type="num">Connected</th></tr></thead>
-                    <tbody>{rows}</tbody>
-                </table>
-            </div>
-        </div>"#,
-        own_loc = own_loc,
-        ring_svg = ring_svg,
-        rows = rows,
-    )
-}
-
-/// Build an SVG ring visualization. Two concentric rings:
-///
-/// - **Outer ring**: peers we're connected to, placed by their
-///   ring-location (Kleinberg topology position).
-/// - **Inner ring**: contracts the governance manager is tracking,
-///   placed by a deterministic hash of the instance id. Colored by
-///   governance state.
-///
-/// Lines (Bezier curves through the interior) go YOU → peer (outer)
-/// and YOU → contract (inner). All curves originate from or terminate
-/// at YOU — a node only sees its own traffic, so anything else would
-/// be fabricated.
-///
-/// Falls back to outer-ring-only rendering when no governance
-/// snapshot is supplied (e.g. tests that pre-date the upgrade).
-fn build_ring_svg(
-    own_location: Option<f64>,
-    peers: &[network_status::PeerSnapshot],
-    governance: Option<&network_status::GovernanceSnapshot>,
-    hosted_contracts: &[network_status::ContractSnapshot],
-) -> String {
-    // Render when there's *something* to show. The historical guard
-    // (at least one peer with a location) still applies.
-    let has_any_location = own_location.is_some() || peers.iter().any(|p| p.location.is_some());
-    if !has_any_location {
-        return String::new();
-    }
-
-    // Larger viewBox than the original 240×240 so the inner ring +
-    // labels have room. Use a viewBox-only sizing (no fixed width)
-    // so the SVG scales with the card.
-    let size: f64 = 480.0;
-    let cx: f64 = size / 2.0;
-    let cy: f64 = size / 2.0;
-    let r_outer: f64 = 195.0;
-    let r_inner: f64 = 120.0;
-
-    let mut svg = format!(
-        "<div class=\"ring-wrap\"><svg viewBox=\"0 0 {size:.0} {size:.0}\" class=\"ring-svg\" preserveAspectRatio=\"xMidYMid meet\">"
-    );
-
-    // Show the inner ring when there are *any* contracts to display
-    // (governance-flagged OR hosted).  Without this the hosted dots
-    // render orphaned on an invisible ring.
-    let has_inner_ring =
-        governance.is_some_and(|g| !g.contracts.is_empty()) || !hosted_contracts.is_empty();
-
-    // === Background rings ===
-    write!(
-        svg,
-        "<circle cx=\"{cx}\" cy=\"{cy}\" r=\"{r_outer}\" fill=\"none\" stroke=\"#363c4a\" stroke-width=\"1\"/>"
-    )
-    .ok();
-    if has_inner_ring {
-        write!(
-            svg,
-            "<circle cx=\"{cx}\" cy=\"{cy}\" r=\"{r_inner}\" fill=\"none\" stroke=\"#363c4a\" stroke-width=\"1\"/>"
-        )
-        .ok();
-    }
-
-    // Helper: location (0.0..1.0) → (x, y) on a ring of given radius.
-    // 0.0 is at the top, increasing clockwise.
-    let loc_to_xy = |loc: f64, r: f64| -> (f64, f64) {
-        let angle = loc * std::f64::consts::TAU - std::f64::consts::FRAC_PI_2;
-        (cx + r * angle.cos(), cy + r * angle.sin())
-    };
-
-    // Curved chord path generator. Quadratic Bezier from (x1,y1) to
-    // (x2,y2) with the control point pulled toward the SVG centre.
-    // `bend_base` is the maximum bend for antipodal points; chord
-    // length scales bend so short hops look flat and long hops arc
-    // through the interior. Matches the nova.locut.us:3133 ring
-    // routing visualization.
-    let curve_path = |x1: f64, y1: f64, x2: f64, y2: f64, bend_base: f64| -> String {
-        let dx = x2 - x1;
-        let dy = y2 - y1;
-        let chord = (dx * dx + dy * dy).sqrt();
-        let ratio = (chord / (2.0 * r_outer)).min(1.0);
-        let bend = bend_base * ratio;
-        let mx = (x1 + x2) / 2.0;
-        let my = (y1 + y2) / 2.0;
-        let pcx = cx + (mx - cx) * (1.0 - bend);
-        let pcy = cy + (my - cy) * (1.0 - bend);
-        format!("M {x1:.1},{y1:.1} Q {pcx:.1},{pcy:.1} {x2:.1},{y2:.1}")
-    };
-
-    // === Ring labels at top ===
-    write!(
-        svg,
-        "<text x=\"{cx}\" y=\"{y:.1}\" text-anchor=\"middle\" fill=\"#6b7280\" font-family=\"monospace\" font-size=\"9\" letter-spacing=\"0.18em\">PEERS</text>",
-        y = cy - r_outer - 8.0,
-    )
-    .ok();
-    if has_inner_ring {
-        write!(
-            svg,
-            "<text x=\"{cx}\" y=\"{y:.1}\" text-anchor=\"middle\" fill=\"#6b7280\" font-family=\"monospace\" font-size=\"9\" letter-spacing=\"0.18em\">CONTRACTS</text>",
-            y = cy - r_inner - 8.0,
-        )
-        .ok();
-    }
-
-    let own_xy = own_location.map(|loc| loc_to_xy(loc, r_outer));
-
-    // === Connection curves: YOU → peers ===
-    // Stroke width scales with total bytes transferred to this peer,
-    // giving a visual "data flow" indication.  Floor at 0.6 so even
-    // idle peers are visible; ceiling at 3.0 for the busiest peer.
-    // Gateways get a warm amber arc, regular peers a teal one.
-    if let Some((ox, oy)) = own_xy {
-        // Find max transfer for relative scaling
-        let max_xfer = peers
-            .iter()
-            .map(|p| p.bytes_sent.saturating_add(p.bytes_received))
-            .max()
-            .unwrap_or(1)
-            .max(1);
-        for p in peers {
-            if let Some(ploc) = p.location {
-                let (px, py) = loc_to_xy(ploc, r_outer);
-                let total = p.bytes_sent.saturating_add(p.bytes_received);
-                let sw = 0.6 + 2.4 * (total as f64 / max_xfer as f64);
-                let (stroke, opacity) = if p.is_gateway {
-                    ("#f0a030", 0.55)
-                } else {
-                    ("#0abab5", 0.45)
-                };
-                let path = curve_path(ox, oy, px, py, 0.55);
-                write!(
-                    svg,
-                    "<path d=\"{path}\" fill=\"none\" stroke=\"{stroke}\" stroke-width=\"{sw:.1}\" stroke-opacity=\"{opacity}\" stroke-linecap=\"round\"/>"
-                )
-                .ok();
-            }
-        }
-    }
-
-    // Deterministic hash → ring location for placing contracts.
-    // The instance id is a 32-byte content hash; we fold the
-    // string form to a u64 and modulo onto [0, 1). Position is
-    // stable across refreshes (same id → same dot location).
-    let hash_to_loc = |s: &str| -> f64 {
-        use std::collections::hash_map::DefaultHasher;
-        use std::hash::{Hash, Hasher};
-        let mut h = DefaultHasher::new();
-        s.hash(&mut h);
-        (h.finish() % 10_000) as f64 / 10_000.0
-    };
-
-    // === Hosted contracts on the inner ring (faint dots) ===
-    //
-    // Before drawing flagged-contract markers, draw ALL hosted
-    // contracts as faint dim dots so the inner ring isn't visually
-    // empty in healthy state. Hosted-contract data comes from the
-    // Subscribed Contracts list (different source from governance,
-    // which only carries flagged entries). A contract that's also
-    // flagged will get a brighter overlay drawn on top below.
-    //
-    // This addresses the "inner ring looks broken" finding from the
-    // UI feedback doc — operators saw the CONTRACTS wordmark with
-    // no dots beneath it and assumed the renderer was unfinished.
-    let governance_ids: std::collections::HashSet<&str> = governance
-        .map(|g| g.contracts.iter().map(|c| c.instance_id.as_str()).collect())
-        .unwrap_or_default();
-    for c in hosted_contracts {
-        // Skip contracts that are flagged — they get a more visible
-        // marker in the flagged-rendering loop below. Drawing them
-        // here would just be overlapped.
-        if governance_ids.contains(c.instance_id.as_str()) {
-            continue;
-        }
-        // Hash on the instance_id so the hosted-dot position matches
-        // the same contract's flagged-dot position (which also hashes
-        // on instance_id) — Codex review caught the previous
-        // key_full vs instance_id mismatch.
-        let loc = hash_to_loc(&c.instance_id);
-        let (kx, ky) = loc_to_xy(loc, r_inner);
-        // Dim teal dot — same brand color as YOU but smaller and
-        // translucent so flagged dots stand out by contrast.
-        write!(
-            svg,
-            "<circle cx=\"{x:.1}\" cy=\"{y:.1}\" r=\"2.5\" fill=\"#43c178\" fill-opacity=\"0.45\"><title>{title}</title></circle>",
-            x = kx,
-            y = ky,
-            title = html_escape(&format!("{} (hosted)", &c.key_short)),
-        )
-        .ok();
-    }
-
-    // === Contract dots (and YOU → flagged-contract curves) ===
-    if let Some(gov) = governance {
-        for c in &gov.contracts {
-            let loc = hash_to_loc(&c.instance_id);
-            let (kx, ky) = loc_to_xy(loc, r_inner);
-            let (fill, glow) = match c.state {
-                network_status::GovernanceStateSnapshot::Normal => ("#43c178", false),
-                network_status::GovernanceStateSnapshot::Borderline => ("#ffb610", false),
-                network_status::GovernanceStateSnapshot::WouldEvict => ("#ff8a3d", true),
-                network_status::GovernanceStateSnapshot::Evicted => ("#ff667a", true),
-                network_status::GovernanceStateSnapshot::Banned => ("#d33682", true),
-            };
-            // Flagged contracts get a curve from YOU into the inner
-            // ring + a small label of the short instance id. Normal
-            // ones get a dot only — they'd otherwise overwhelm the
-            // visual with curves.
-            let is_flagged = !matches!(c.state, network_status::GovernanceStateSnapshot::Normal);
-            if is_flagged {
-                if let Some((ox, oy)) = own_xy {
-                    let path = curve_path(ox, oy, kx, ky, 0.6);
-                    write!(
-                        svg,
-                        "<path d=\"{path}\" fill=\"none\" stroke=\"{fill}\" stroke-width=\"1.4\" stroke-opacity=\"0.7\"/>"
-                    )
-                    .ok();
-                }
-            }
-            // Use a more distinctive shape for flagged (rect) than
-            // normal (smaller, dimmer dot) so glance-scanning the
-            // ring surfaces flagged contracts first.
-            let size_px = if is_flagged { 6.0 } else { 3.0 };
-            let opacity = if is_flagged { "1.0" } else { "0.55" };
-            let glow_attr = if glow {
-                "filter=\"drop-shadow(0 0 3px currentColor)\""
-            } else {
-                ""
-            };
-            write!(
-                svg,
-                "<rect x=\"{x:.1}\" y=\"{y:.1}\" width=\"{size_px:.1}\" height=\"{size_px:.1}\" fill=\"{fill}\" fill-opacity=\"{opacity}\" {glow_attr}><title>{title}</title></rect>",
-                x = kx - size_px / 2.0,
-                y = ky - size_px / 2.0,
-                title = html_escape(&format!("{} ({:?})", c.instance_id_short, c.state)),
-            )
-            .ok();
-            // Short label next to flagged contracts so the ring view
-            // matches the table without clicking.
-            if is_flagged {
-                let label_loc_r = r_inner - 12.0;
-                let (lx, ly) = loc_to_xy(loc, label_loc_r);
-                write!(
-                    svg,
-                    "<text x=\"{lx:.1}\" y=\"{ly:.1}\" text-anchor=\"middle\" dominant-baseline=\"middle\" fill=\"{fill}\" font-family=\"monospace\" font-size=\"9\" font-weight=\"500\">{label}</text>",
-                    label = html_escape(&c.instance_id_short),
-                )
-                .ok();
-            }
-        }
-    }
-
-    // === Peer dots on the outer ring ===
-    for p in peers {
-        if let Some(loc) = p.location {
-            let (px, py) = loc_to_xy(loc, r_outer);
-            let fill = if p.is_gateway { "#ffb610" } else { "#66d9ff" };
-            let kind = if p.is_gateway { "Gateway" } else { "Peer" };
-            let addr = p.address.to_string();
-            let title = format!("{kind} {addr} (loc {loc:.4})");
-            write!(
-                svg,
-                "<a href=\"/peer/{href}\" class=\"ring-peer-link\"><title>{title}</title><circle cx=\"{px:.1}\" cy=\"{py:.1}\" r=\"4\" fill=\"{fill}\"/></a>",
-                href = html_escape(&addr),
-                title = html_escape(&title),
-            )
-            .ok();
-        }
-    }
-
-    // === YOU marker — drawn last so it sits above everything ===
-    if let Some(own_loc) = own_location {
-        let (ox, oy) = loc_to_xy(own_loc, r_outer);
-        write!(
-            svg,
-            "<g class=\"ring-self\"><title>You (loc {own_loc:.4})</title><circle cx=\"{ox:.1}\" cy=\"{oy:.1}\" r=\"6\" fill=\"#43c178\" stroke=\"#ebecf0\" stroke-width=\"1.5\" class=\"you-dot\"/><text x=\"{lx:.1}\" y=\"{ly:.1}\" text-anchor=\"middle\" fill=\"#43c178\" font-family=\"monospace\" font-size=\"9\" font-weight=\"500\" letter-spacing=\"0.05em\">YOU</text></g>",
-            lx = ox,
-            ly = oy + 18.0,
-        )
-        .ok();
-    }
-
-    svg.push_str("</svg>");
-
-    // Legend below the ring.
-    svg.push_str(concat!(
-        "<div class=\"ring-legend\">",
-        "<span class=\"ring-key\"><span class=\"ring-dot ring-dot-self\"></span> You</span>",
-        "<span class=\"ring-key\"><span class=\"ring-dot ring-dot-peer\"></span> Peer</span>",
-        "<span class=\"ring-key\"><span class=\"ring-dot ring-dot-gw\"></span> Gateway</span>",
-        "<span class=\"ring-key\"><span class=\"ring-dot ring-dot-contract-normal\"></span> Hosted</span>",
-        "<span class=\"ring-key\"><span class=\"ring-dot ring-dot-contract-flagged\"></span> Flagged</span>",
-        "</div></div>",
-    ));
-
-    svg
-}
-
-/// Format the governance card's "last evaluated" footer from the number
-/// of seconds since the reaper last ticked. `format_ago` already appends
-/// " ago" (or returns "just now"), so the template must NOT add a second
-/// "ago" — doing so rendered "Last evaluated 18s ago ago".
-fn format_last_evaluated(secs: u64) -> String {
-    format!("Last evaluated {}", format_ago(secs))
-}
-
-/// Build the governance card. Reads `snap.governance` (sourced from
-/// `Ring::dashboard_governance_snapshot` → `GovernanceManager`). Every
-/// field rendered here came from the back-end's computation —
-/// nothing is invented at render time.
-///
-/// Layout follows the prototype: verdict block on the left (big
-/// number + headline), mini-strip with network-norms on the right,
-/// then the per-contract table below. The histogram and ring inner-
-/// ring renderings are deliberately separate commits — this commit
-/// proves the data path end-to-end with the simplest visualisations.
-fn build_governance_card(snap: &Option<network_status::NetworkStatusSnapshot>) -> String {
-    let Some(snap) = snap else {
-        return String::new();
-    };
-    let g = &snap.governance;
-
-    let mode_txt = match g.mode {
-        network_status::GovernanceModeSnapshot::Off => "off",
-        network_status::GovernanceModeSnapshot::DryRun => "dry-run",
-        network_status::GovernanceModeSnapshot::Enforce => "enforce",
-    };
-
-    let last_tick_footer = match g.last_tick_at {
-        Some(at) => {
-            // Compute "Ns ago" using TOKIO instant arithmetic via
-            // a wall-clock comparison. The snapshot was built
-            // moments ago so we can approximate "now" inline.
-            let now = tokio::time::Instant::now();
-            let secs = now.saturating_duration_since(at).as_secs();
-            format_last_evaluated(secs)
-        }
-        None => "Reaper has not yet ticked".to_string(),
-    };
-
-    // Empty state: no FLAGGED contracts. Render the structural
-    // skeleton (mode pill, 5-tile mini-strip with em-dashes if data
-    // is unavailable, observed/required progress) rather than just a
-    // paragraph — teaches operators what data will appear and what
-    // mode is active. The previous empty state hid this and made
-    // an operator think the dashboard was half-implemented.
-    if g.contracts.is_empty() {
-        let observed = g.observed_count;
-        let needed = g.min_samples;
-        // When governance is Off (the default — see `GovernanceConfig`),
-        // no contract is ever scored, so the ramp-up / "scoring activates"
-        // copy below would be misleading on every default node. Render a
-        // disabled message instead.
-        let is_off = matches!(g.mode, network_status::GovernanceModeSnapshot::Off);
-        // Tiny pluralization helper so the user-facing messages
-        // don't read "1 contracts" — Codex review nit.
-        let plural = |n: usize| if n == 1 { "contract" } else { "contracts" };
-        let progress_msg = if is_off {
-            "Governance is off: contracts are not scored, and nothing is evicted.".to_string()
-        } else if needed == 0 {
-            "Governance manager is not yet wired.".to_string()
-        } else if observed == 0 {
-            format!(
-                "No contracts observed yet. Scoring activates after {needed} {n_word} \
-                 have accumulated cost.",
-                n_word = plural(needed),
-            )
-        } else if observed < needed {
-            let remaining = needed - observed;
-            let verb = if remaining == 1 {
-                "accumulates"
-            } else {
-                "accumulate"
-            };
-            format!(
-                "Observed {observed} / {needed} {n_word} needed for statistical scoring. \
-                 Scoring activates once {remaining} more {r_word} {verb} cost.",
-                n_word = plural(needed),
-                r_word = plural(remaining),
-            )
-        } else {
-            // Enough samples observed but none flagged — that's the
-            // healthy steady state.
-            format!(
-                "All {observed} tracked {n_word} within normal range. \
-                 (Scored against the network's own observed distribution.)",
-                n_word = plural(observed),
-            )
-        };
-        let verdict_main = if is_off {
-            r#"<div class="verdict-num">—</div>
-                   <div class="verdict-headline">Governance off</div>
-                   <div class="verdict-detail">Contracts are not scored or evicted.</div>"#
-                .to_string()
-        } else if observed >= needed {
-            format!(
-                r#"<div class="verdict-num">✓</div>
-                   <div class="verdict-headline">{observed} contracts within normal range</div>
-                   <div class="verdict-detail">No flags raised.</div>"#
-            )
-        } else {
-            format!(
-                r#"<div class="verdict-num">{observed}<span class="verdict-num-denom">/{needed}</span></div>
-                   <div class="verdict-headline">contracts observed</div>
-                   <div class="verdict-detail">Scoring activates at {needed}.</div>"#
-            )
-        };
-
-        // 5-tile skeleton — render even with no data, using em-dashes
-        // for missing values. This shows operators what fields will
-        // populate as the reaper ticks.
-        let median_txt = g
-            .norms
-            .median_log_ratio
-            .map(|v| format!("{:.2}", v))
-            .unwrap_or_else(|| "—".to_string());
-        let mad_txt = g
-            .norms
-            .mad
-            .map(|v| format!("{:.2}", v))
-            .unwrap_or_else(|| "—".to_string());
-        let threshold_txt = g
-            .norms
-            .threshold
-            .map(|v| format!("{:.2}", v))
-            .unwrap_or_else(|| "—".to_string());
-        let sample_size_txt = if g.norms.sample_size == 0 {
-            "—".to_string()
-        } else {
-            g.norms.sample_size.to_string()
-        };
-
-        return format!(
-            r##"<div class="card">
-                <div class="card-header"><h2>Contract Governance</h2><span class="g-mode g-mode-{mode}">{mode}</span></div>
-                <div class="g-verdict-row">
-                    <div class="g-verdict verdict-ok">{verdict_main}</div>
-                    <div class="g-norms">
-                        <div class="g-norm"><div class="g-norm-label">Tracked</div><div class="g-norm-value">{observed}</div></div>
-                        <div class="g-norm"><div class="g-norm-label">Sample size</div><div class="g-norm-value">{sample_size}</div></div>
-                        <div class="g-norm"><div class="g-norm-label">Median log-ratio</div><div class="g-norm-value">{median}</div></div>
-                        <div class="g-norm"><div class="g-norm-label">MAD spread</div><div class="g-norm-value">{mad}</div></div>
-                        <div class="g-norm"><div class="g-norm-label">Eviction threshold</div><div class="g-norm-value">{threshold}</div></div>
-                    </div>
-                </div>
-                <p class="empty" style="margin: 0.6rem 0.9rem 0.2rem; font-size: 0.9rem;">{progress}</p>
-                <p class="empty" style="margin: 0 0.9rem 0.6rem; font-size: 0.78rem; color: var(--text-muted, #888);">{tick_footer}</p>
-            </div>"##,
-            mode = mode_txt,
-            verdict_main = verdict_main,
-            observed = observed,
-            sample_size = sample_size_txt,
-            median = median_txt,
-            mad = mad_txt,
-            threshold = threshold_txt,
-            progress = progress_msg,
-            tick_footer = last_tick_footer,
-        );
-    }
-
-    // Verdict counts. Mirror state-snapshot enum → string.
-    let mut counts = [0u32; 5];
-    for c in &g.contracts {
-        let idx = match c.state {
-            network_status::GovernanceStateSnapshot::Normal => 0,
-            network_status::GovernanceStateSnapshot::Borderline => 1,
-            network_status::GovernanceStateSnapshot::WouldEvict => 2,
-            network_status::GovernanceStateSnapshot::Evicted => 3,
-            network_status::GovernanceStateSnapshot::Banned => 4,
-        };
-        counts[idx] = counts[idx].saturating_add(1);
-    }
-    let borderline = counts[1];
-    let would_evict = counts[2];
-    let evicted = counts[3];
-    let banned = counts[4];
-    let flagged = borderline + would_evict + evicted + banned;
-    let total = g.contracts.len();
-
-    let verdict_class = if flagged == 0 {
-        "verdict-ok"
-    } else {
-        "verdict-alert"
-    };
-    let verdict_main = if flagged == 0 {
-        format!(
-            r#"<div class="verdict-num">✓</div>
-               <div class="verdict-headline">All {total} contracts within normal range</div>"#,
-        )
-    } else {
-        let mut detail_parts: Vec<String> = Vec::new();
-        if would_evict > 0 {
-            detail_parts.push(format!(
-                r#"<span class="sw sw-wouldevict"></span>{would_evict} would be evicted"#
-            ));
-        }
-        if borderline > 0 {
-            detail_parts.push(format!(
-                r#"<span class="sw sw-borderline"></span>{borderline} borderline"#
-            ));
-        }
-        if evicted > 0 {
-            detail_parts.push(format!(
-                r#"<span class="sw sw-evicted"></span>{evicted} evicted"#
-            ));
-        }
-        if banned > 0 {
-            detail_parts.push(format!(
-                r#"<span class="sw sw-banned"></span>{banned} banned"#
-            ));
-        }
-        format!(
-            r#"<div class="verdict-num">{flagged}</div>
-               <div class="verdict-headline">contracts flagged on this node</div>
-               <div class="verdict-detail">{detail}</div>"#,
-            detail = detail_parts.join(" &nbsp;·&nbsp; "),
-        )
-    };
-
-    // Network-norms mini-strip — sourced from the last reaper-tick
-    // result. Empty if no tick has run yet (cold start; sample size
-    // didn't reach min_samples; MAD collapsed).
-    let median_txt = g
-        .norms
-        .median_log_ratio
-        .map(|v| format!("{:.2}", v))
-        .unwrap_or_else(|| "—".to_string());
-    let mad_txt = g
-        .norms
-        .mad
-        .map(|v| format!("{:.2}", v))
-        .unwrap_or_else(|| "—".to_string());
-    let threshold_txt = g
-        .norms
-        .threshold
-        .map(|v| format!("{:.2}", v))
-        .unwrap_or_else(|| "—".to_string());
-    let sample_size_txt = g.norms.sample_size.to_string();
-    // `mode_txt` and `last_tick_footer` are in scope from the top
-    // of this function (defined once, reused in both empty and
-    // populated branches).
-
-    // Per-contract table — flagged-only by default; an "all" link
-    // could come later. Honest principle: this table reflects the
-    // governance manager's `iter_scores()`, nothing else.
-    let mut rows = String::new();
-    let mut shown_count = 0;
-    // Sort: most-flagged first (Banned > Evicted > WouldEvict >
-    // Borderline > Normal), then by highest log-ratio descending so
-    // the worst offenders sit at the top.
-    let mut sorted: Vec<&network_status::ContractGovernanceEntry> = g.contracts.iter().collect();
-    sorted.sort_by(|a, b| {
-        let rank = |s: network_status::GovernanceStateSnapshot| match s {
-            network_status::GovernanceStateSnapshot::Banned => 0,
-            network_status::GovernanceStateSnapshot::Evicted => 1,
-            network_status::GovernanceStateSnapshot::WouldEvict => 2,
-            network_status::GovernanceStateSnapshot::Borderline => 3,
-            network_status::GovernanceStateSnapshot::Normal => 4,
-        };
-        rank(a.state).cmp(&rank(b.state)).then_with(|| {
-            b.log_ratio
-                .partial_cmp(&a.log_ratio)
-                .unwrap_or(std::cmp::Ordering::Equal)
-        })
-    });
-    for c in sorted.iter() {
-        // Default: only show flagged contracts. An "all" toggle could
-        // come later; for now keeping the table digestible.
-        if matches!(c.state, network_status::GovernanceStateSnapshot::Normal) {
-            continue;
-        }
-        shown_count += 1;
-        let state_label = match c.state {
-            network_status::GovernanceStateSnapshot::Normal => "normal",
-            network_status::GovernanceStateSnapshot::Borderline => "borderline",
-            network_status::GovernanceStateSnapshot::WouldEvict => "would evict",
-            network_status::GovernanceStateSnapshot::Evicted => "evicted",
-            network_status::GovernanceStateSnapshot::Banned => "banned",
-        };
-        let state_class = match c.state {
-            network_status::GovernanceStateSnapshot::Normal => "g-normal",
-            network_status::GovernanceStateSnapshot::Borderline => "g-borderline",
-            network_status::GovernanceStateSnapshot::WouldEvict => "g-wouldevict",
-            network_status::GovernanceStateSnapshot::Evicted => "g-evicted",
-            network_status::GovernanceStateSnapshot::Banned => "g-banned",
-        };
-        let log_ratio_txt = c
-            .log_ratio
-            .map(|v| format!("{:+.2}", v))
-            .unwrap_or_else(|| "—".to_string());
-        let age = format_ago(c.age_secs);
-        let state_rank = match c.state {
-            network_status::GovernanceStateSnapshot::Banned => 0u8,
-            network_status::GovernanceStateSnapshot::Evicted => 1,
-            network_status::GovernanceStateSnapshot::WouldEvict => 2,
-            network_status::GovernanceStateSnapshot::Borderline => 3,
-            network_status::GovernanceStateSnapshot::Normal => 4,
-        };
-        let log_ratio_sort = c.log_ratio.map(|v| format!("{v:.6}")).unwrap_or_default();
-        rows.push_str(&format!(
-            r#"<tr><td title="{full}" data-sort="{full}"><code>{short}</code><button type="button" class="copy-key" data-copy="{full}" title="Copy contract key" aria-label="Copy contract key">⧉</button></td><td data-sort="{state_rank}"><span class="g-badge {state_class}">{state_label}</span></td><td class="right" data-sort="{log_ratio_sort}">{log_ratio}</td><td class="right" data-sort="{cost:.6}">{cost:.2}</td><td class="right" data-sort="{benefit:.6}">{benefit:.2}</td><td class="right" data-sort="{age_secs}">{age}</td></tr>"#,
-            full = html_escape(&c.instance_id),
-            short = html_escape(&c.instance_id_short),
-            state_class = state_class,
-            state_label = state_label,
-            state_rank = state_rank,
-            log_ratio = log_ratio_txt,
-            log_ratio_sort = log_ratio_sort,
-            cost = c.cost_used,
-            benefit = c.benefit_score,
-            age = age,
-            age_secs = c.age_secs,
-        ));
-    }
-    if shown_count == 0 {
-        rows = r#"<tr class="sort-disabled"><td colspan="6" class="empty" style="padding: 0.5rem 0.9rem">All contracts within normal range.</td></tr>"#.to_string();
-    }
-
-    let tracked_total = g.observed_count.max(total);
-    format!(
-        r##"<div class="card">
-            <div class="card-header"><h2>Contract Governance</h2><span class="g-mode g-mode-{mode}">{mode}</span></div>
-            <div class="g-verdict-row">
-                <div class="g-verdict {verdict_class}">{verdict_main}</div>
-                <div class="g-norms">
-                    <div class="g-norm"><div class="g-norm-label">Tracked</div><div class="g-norm-value">{tracked_total}</div></div>
-                    <div class="g-norm"><div class="g-norm-label">Sample size</div><div class="g-norm-value">{sample_size}</div></div>
-                    <div class="g-norm"><div class="g-norm-label">Median log-ratio</div><div class="g-norm-value">{median}</div></div>
-                    <div class="g-norm"><div class="g-norm-label">MAD spread</div><div class="g-norm-value">{mad}</div></div>
-                    <div class="g-norm"><div class="g-norm-label">Eviction threshold</div><div class="g-norm-value">{threshold}</div></div>
-                </div>
-            </div>
-            <div class="table-wrap">
-                <table class="sortable" data-table-id="governance">
-                    <thead><tr><th data-sort-type="text">Contract</th><th data-sort-type="num">State</th><th class="right" data-sort-type="num">log-ratio</th><th class="right" data-sort-type="num">Cost</th><th class="right" data-sort-type="num">Benefit</th><th class="right" data-sort-type="num">Age</th></tr></thead>
-                    <tbody>{rows}</tbody>
-                </table>
-            </div>
-            <p class="empty" style="margin: 0.4rem 0.9rem 0.6rem; font-size: 0.78rem; color: var(--text-muted, #888);">{tick_footer}</p>
-        </div>"##,
-        mode = mode_txt,
-        verdict_class = verdict_class,
-        verdict_main = verdict_main,
-        tracked_total = tracked_total,
-        tick_footer = last_tick_footer,
-        sample_size = sample_size_txt,
-        median = median_txt,
-        mad = mad_txt,
-        threshold = threshold_txt,
-        rows = rows,
-    )
-}
-
-/// Contract ban-list card (#4302). Surfaces the canonical
-/// `Ring::contract_ban_list` so operators can see whether the Phase 7
-/// hardening mechanism is catching abusers or sitting idle.
-///
-/// Two concrete asks from the issue: a count tile ("N contracts on ban
-/// list") and a per-entry list (key, reason, time remaining). The
-/// governance-state-machine drill-down is deferred (the ban list stores
-/// only the current entry, not the transition history that led to it).
-///
-/// The card is rendered whenever a snapshot exists — including when the
-/// list is empty — so an idle-but-active mechanism is distinguishable
-/// from one that isn't wired. Every value here comes from the ban list's
-/// own accessors via the provider closure; nothing is invented at render
-/// time.
-fn build_ban_list_card(snap: &Option<network_status::NetworkStatusSnapshot>) -> String {
-    let Some(snap) = snap else {
-        return String::new();
-    };
-    let b = &snap.ban_list;
-
-    let plural = |n: usize| if n == 1 { "contract" } else { "contracts" };
-
-    // Capacity-rejection note: only shown when non-zero, so the common
-    // case stays uncluttered. A non-zero value is the operator's signal
-    // that the bounded list (MAX_BANNED_CONTRACTS) is overflowing.
-    let capacity_note = if b.capacity_rejected_total > 0 {
-        format!(
-            r#"<p class="empty" style="margin: 0 0.9rem 0.6rem; font-size: 0.82rem; color: var(--danger, #c0392b);">{n} ban{s} rejected — list at capacity.</p>"#,
-            n = b.capacity_rejected_total,
-            s = if b.capacity_rejected_total == 1 {
-                ""
-            } else {
-                "s"
-            },
-        )
-    } else {
-        String::new()
-    };
-
-    let body = if b.entries.is_empty() {
-        r#"<p class="empty" style="margin: 0.6rem 0.9rem;">No contracts banned. The mechanism is active and currently idle.</p>"#
-            .to_string()
-    } else {
-        let mut rows = String::new();
-        for e in &b.entries {
-            let reason_txt = match e.reason {
-                network_status::BanReasonSnapshot::AutoMad => "auto (governance)",
-                network_status::BanReasonSnapshot::Operator => "operator",
-            };
-            let remaining = if e.expires_in_secs == 0 {
-                "lifting".to_string()
-            } else {
-                format!("{} left", format_duration(e.expires_in_secs))
-            };
-            // html_escape the contract id even though instance ids are
-            // base58 (no HTML metacharacters today): defense-in-depth so
-            // the row stays safe if the id format ever changes.
-            write!(
-                rows,
-                r#"<tr><td class="mono">{id}</td><td>{reason}</td><td>{remaining}</td></tr>"#,
-                id = html_escape(&e.instance_id),
-                reason = reason_txt,
-                remaining = html_escape(&remaining),
-            )
-            .ok();
-        }
-        format!(
-            r#"<table class="data-table">
-                <thead><tr><th>Contract</th><th>Reason</th><th>Expires</th></tr></thead>
-                <tbody>{rows}</tbody>
-            </table>"#
-        )
-    };
-
-    format!(
-        r##"<div class="card">
-            <div class="card-header"><h2>Contract Ban List</h2></div>
-            <div class="g-verdict-row">
-                <div class="g-norms">
-                    <div class="g-norm"><div class="g-norm-label">On ban list</div><div class="g-norm-value">{count}</div></div>
-                </div>
-                <p class="empty" style="margin: 0; padding: 0.4rem 0.9rem; font-size: 0.9rem;">{count} {n_word} currently banned at this node.</p>
-            </div>
-            {capacity_note}
-            {body}
-        </div>"##,
-        count = b.count,
-        n_word = plural(b.count),
-        capacity_note = capacity_note,
-        body = body,
-    )
-}
-
-fn build_contracts_card(snap: &Option<network_status::NetworkStatusSnapshot>) -> String {
-    let Some(snap) = snap else {
-        return String::new();
-    };
-    if snap.contracts.is_empty() {
-        if snap.open_connections > 0 {
-            return r#"<div class="card">
-                <h2>Subscribed Contracts</h2>
-                <p class="empty">No active subscriptions</p>
-            </div>"#
-                .to_string();
-        }
-        return String::new();
-    }
-
-    let state_by_id = &snap.governance.state_by_id;
-    let mut rows = String::new();
-    for c in &snap.contracts {
-        let last_update = c
-            .last_updated_secs
-            .map(format_ago)
-            .unwrap_or_else(|| "—".to_string());
-        let last_update_sort = c.last_updated_secs.unwrap_or(u64::MAX);
-        // Governance state cell. A contract may appear in the
-        // Subscribed Contracts table without being flagged by
-        // governance — in that case the state isn't in `state_by_id`
-        // (the snapshot only carries flagged contracts) and we
-        // render it as "ok". Absence from the table specifically
-        // means "not flagged"; we trust the back-end's `iter_flagged_
-        // scores` filter.
-        let (gov_class, gov_label, gov_sort) = match state_by_id.get(&c.instance_id) {
-            None => ("gov-ok", "ok", 0u8),
-            Some(network_status::GovernanceStateSnapshot::Normal) => ("gov-ok", "ok", 0),
-            Some(network_status::GovernanceStateSnapshot::Borderline) => {
-                ("gov-borderline", "borderline", 1)
-            }
-            Some(network_status::GovernanceStateSnapshot::WouldEvict) => {
-                ("gov-wouldevict", "would evict", 2)
-            }
-            Some(network_status::GovernanceStateSnapshot::Evicted) => ("gov-evicted", "evicted", 3),
-            Some(network_status::GovernanceStateSnapshot::Banned) => ("gov-banned", "banned", 4),
-        };
-        rows.push_str(&format!(
-            r#"<tr><td title="{full}" data-sort="{full}"><code>{short}</code><button type="button" class="copy-key" data-copy="{full}" title="Copy contract key" aria-label="Copy contract key">⧉</button></td><td data-sort="{gov_sort}"><span class="gov-pill {gov_class}">{gov_label}</span></td><td data-sort="{sub_secs}">{subscribed}</td><td data-sort="{last_sort}">{last_update}</td></tr>"#,
-            full = html_escape(&c.key_full),
-            short = html_escape(&c.key_short),
-            gov_sort = gov_sort,
-            gov_class = gov_class,
-            gov_label = gov_label,
-            sub_secs = c.subscribed_secs,
-            subscribed = format_ago(c.subscribed_secs),
-            last_sort = last_update_sort,
-            last_update = last_update,
-        ));
-    }
-
-    format!(
-        r#"<div class="card">
-            <h2>Subscribed Contracts</h2>
-            <div class="table-wrap">
-                <table class="sortable" data-table-id="contracts">
-                    <thead><tr><th data-sort-type="text">Contract</th><th data-sort-type="num">Gov</th><th data-sort-type="num">Subscribed</th><th data-sort-type="num">Last Update</th></tr></thead>
-                    <tbody>{rows}</tbody>
-                </table>
-            </div>
-        </div>"#,
-        rows = rows,
-    )
-}
-
-fn build_ops_card(snap: &Option<network_status::NetworkStatusSnapshot>) -> String {
-    let Some(snap) = snap else {
-        return String::new();
-    };
-    let ops = &snap.op_stats;
-    if ops.total() == 0 && snap.open_connections == 0 {
-        return String::new();
-    }
-
-    fn op_cell(name: &str, ok: u32, fail: u32) -> String {
-        format!(
-            r#"<div class="op-cell">
-                <div class="op-name">{name}</div>
-                <div><span class="op-ok">{ok}</span> <span class="op-fail">{fail}</span></div>
-            </div>"#,
-            name = name,
-            ok = ok,
-            fail = fail,
-        )
-    }
-
-    // UPDATE cell: show received broadcast count (single number) since
-    // subscription-streamed updates are push-based and don't have success/failure.
-    // If there are also routed updates (with success/fail), show both.
-    let update_cell = {
-        let routed = ops.updates.0 + ops.updates.1;
-        let received = ops.updates_received;
-        if routed > 0 {
-            // Both routed and received
-            format!(
-                r#"<div class="op-cell">
-                    <div class="op-name">UPDATE</div>
-                    <div><span class="op-ok">{ok}</span> <span class="op-fail">{fail}</span></div>
-                    <div class="op-received">{recv} received</div>
-                </div>"#,
-                ok = ops.updates.0,
-                fail = ops.updates.1,
-                recv = received,
-            )
-        } else {
-            // Only received (common for subscriber nodes)
-            format!(
-                r#"<div class="op-cell">
-                    <div class="op-name">UPDATE</div>
-                    <div class="op-count">{recv}</div>
-                </div>"#,
-                recv = received,
-            )
-        }
-    };
-
-    format!(
-        r#"<div class="card">
-            <h2>Operations</h2>
-            <div class="op-grid">
-                {get}{put}{update}{subscribe}
-            </div>
-        </div>"#,
-        get = op_cell("GET", ops.gets.0, ops.gets.1),
-        put = op_cell("PUT", ops.puts.0, ops.puts.1),
-        update = update_cell,
-        subscribe = {
-            // Show active subscription count as primary metric since the cumulative
-            // operation count includes periodic lease renewals (every 2 min per contract)
-            // which inflates the number and confuses users.
-            let active = snap.contracts.len() as u32;
-            let total_ops = ops.subscribes.0.saturating_add(ops.subscribes.1);
-            if total_ops > 0 {
-                format!(
-                    r#"<div class="op-cell">
-                        <div class="op-name">SUBSCRIBE</div>
-                        <div class="op-count">{active} active</div>
-                        <div class="op-received">{total_ops} ops</div>
-                    </div>"#,
-                )
-            } else {
-                format!(
-                    r#"<div class="op-cell">
-                        <div class="op-name">SUBSCRIBE</div>
-                        <div class="op-count">{active} active</div>
-                    </div>"#,
-                )
-            }
-        },
-    )
-}
-
-const CSS: &str = r##"
-/* ── CSS Variables (dark mode default) ── */
-:root {
-    --bg-primary: #0c0d0f;
-    --bg-secondary: #141518;
-    --bg-tertiary: #1a1c20;
-    --bg-panel: rgba(20, 21, 24, 0.85);
-    --border-color: rgba(64, 66, 72, 0.35);
-    --text-primary: #edeeef;
-    --text-secondary: #94969a;
-    --text-muted: #585a5e;
-    --accent-primary: #0abab5;
-    --accent-light: #5eead4;
-    --accent-dark: #0d9488;
-    --font-mono: 'Iosevka', 'JetBrains Mono', 'Cascadia Code', 'Fira Code', 'SF Mono', monospace;
-    --font-sans: 'Space Grotesk', 'DM Sans', system-ui, -apple-system, sans-serif;
-}
-/* Light mode overrides */
-[data-theme="light"] {
-    --bg-primary: #f7f5f2;
-    --bg-secondary: #eeebe6;
-    --bg-tertiary: #e5e0d9;
-    --bg-panel: rgba(255, 255, 255, 0.88);
-    --border-color: rgba(180, 175, 168, 0.4);
-    --text-primary: #1a1816;
-    --text-secondary: #5c5955;
-    --text-muted: #8c8985;
-}
-* { box-sizing: border-box; margin: 0; padding: 0; }
-body {
-    font-family: var(--font-sans);
-    background: var(--bg-primary);
-    color: var(--text-primary);
-    line-height: 1.5;
-}
-/* Subtle warm background with dot-grid texture */
-body::before {
-    content: '';
-    position: fixed;
-    top: 0; left: 0; right: 0; bottom: 0;
-    background:
-        radial-gradient(ellipse at 30% 20%, rgba(224, 220, 210, 0.03) 0%, transparent 55%),
-        radial-gradient(ellipse at 70% 60%, rgba(10, 186, 181, 0.04) 0%, transparent 50%);
-    pointer-events: none;
-    z-index: 0;
-}
-body::after {
-    content: '';
-    position: fixed;
-    top: 0; left: 0; right: 0; bottom: 0;
-    background-image: radial-gradient(rgba(255,255,255,0.015) 1px, transparent 1px);
-    background-size: 24px 24px;
-    pointer-events: none;
-    z-index: 0;
-    opacity: 0.6;
-}
-[data-theme="light"] body::before {
-    background:
-        radial-gradient(ellipse at 30% 20%, rgba(180, 175, 165, 0.06) 0%, transparent 55%),
-        radial-gradient(ellipse at 70% 60%, rgba(10, 186, 181, 0.04) 0%, transparent 50%);
-}
-[data-theme="light"] body::after {
-    background-image: radial-gradient(rgba(0,0,0,0.025) 1px, transparent 1px);
-    opacity: 0.5;
-}
-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0.75rem 1.5rem;
-    background: var(--bg-panel);
-    border-bottom: 1px solid var(--border-color);
-    backdrop-filter: blur(10px);
-    position: relative;
-    z-index: 1;
-}
-.header-left {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-}
-.header-right {
-    display: flex;
-    align-items: center;
-    gap: 0.75rem;
-    color: var(--text-secondary);
-    font-size: 0.9rem;
-}
-.logo { width: 28px; height: 28px; }
-.header-title {
-    font-family: var(--font-mono);
-    font-weight: 600;
-    font-size: 1.1rem;
-    letter-spacing: -0.02em;
-    background: linear-gradient(135deg, var(--accent-light) 0%, var(--accent-primary) 50%, var(--accent-dark) 100%);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-}
-.header-scope {
-    font-family: var(--font-mono);
-    font-size: 0.75rem;
-    font-weight: 500;
-    color: var(--accent-light);
-    background: rgba(94, 234, 212, 0.1);
-    padding: 0.15rem 0.5rem;
-    border-radius: 4px;
-    border: 1px solid rgba(94, 234, 212, 0.2);
-    text-transform: uppercase;
-    letter-spacing: 0.05em;
-}
-.pub-key-label {
-    font-size: 0.65rem;
-    color: var(--text-muted);
-    text-transform: uppercase;
-    font-family: var(--font-mono);
-    margin-right: 0.3rem;
-}
-.pub-key {
-    font-family: var(--font-mono);
-    font-size: 0.7rem;
-    color: var(--text-secondary);
-    background: rgba(139, 148, 158, 0.08);
-    padding: 0.15rem 0.5rem;
-    border-radius: 4px 0 0 4px;
-    border: 1px solid var(--border-color);
-    max-width: 24ch;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    cursor: pointer;
-}
-.pub-key:hover { color: var(--accent-light); }
-.copy-btn {
-    font-family: var(--font-mono);
-    font-size: 0.7rem;
-    color: var(--text-secondary);
-    background: rgba(139, 148, 158, 0.08);
-    border: 1px solid var(--border-color);
-    border-left: none;
-    border-radius: 0 4px 4px 0;
-    padding: 0.15rem 0.4rem;
-    cursor: pointer;
-    line-height: 1;
-}
-.copy-btn:hover { color: var(--accent-light); background: rgba(94, 234, 212, 0.1); }
-.badge {
-    background: rgba(10, 186, 181, 0.15);
-    color: var(--accent-light);
-    padding: 0.15rem 0.5rem;
-    border-radius: 12px;
-    font-size: 0.75rem;
-    font-weight: 500;
-    font-family: var(--font-mono);
-}
-.uptime { font-family: var(--font-mono); font-size: 0.85rem; }
-main {
-    max-width: 800px;
-    margin: 1.5rem auto;
-    padding: 0 1rem;
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    position: relative;
-    z-index: 1;
-}
-.card {
-    background: var(--bg-panel);
-    border: 1px solid var(--border-color);
-    border-radius: 12px;
-    padding: 1rem 1.25rem;
-    backdrop-filter: blur(10px);
-    box-shadow: 0 1px 3px rgba(0,0,0,0.12);
-}
-.card-muted { background: var(--bg-secondary); }
-.card h2 {
-    font-size: 0.85rem;
-    color: var(--text-secondary);
-    text-transform: uppercase;
-    letter-spacing: 0.05em;
-    margin-bottom: 0.75rem;
-    font-family: var(--font-mono);
-    font-weight: 500;
-}
-.card-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: baseline;
-    margin-bottom: 0.75rem;
-}
-.card-header h2 { margin-bottom: 0; }
-.own-loc {
-    font-size: 0.8rem;
-    color: var(--text-secondary);
-    font-family: var(--font-mono);
-}
-.status-row {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    font-size: 1rem;
-    font-weight: 500;
-    margin-bottom: 0.5rem;
-}
-.dot {
-    width: 10px;
-    height: 10px;
-    border-radius: 50%;
-    display: inline-block;
-    flex-shrink: 0;
-}
-.dot-green { background: #34d399; box-shadow: 0 0 8px rgba(52, 211, 153, 0.5); }
-.dot-yellow { background: #fbbf24; box-shadow: 0 0 8px rgba(251, 191, 36, 0.5); }
-.spinner {
-    width: 20px;
-    height: 20px;
-    border: 2px solid var(--border-color);
-    border-top-color: var(--accent-primary);
-    border-radius: 50%;
-    animation: spin 1s linear infinite;
-    margin: 0.5rem 0;
-}
-@keyframes spin { to { transform: rotate(360deg); } }
-@keyframes peer-fade-in {
-    from { opacity: 0; transform: translateY(-6px); }
-    to   { opacity: 1; transform: translateY(0); }
-}
-@keyframes you-pulse {
-    0%, 100% { r: 6; }
-    50%      { r: 8.5; }
-}
-.warning {
-    background: rgba(251, 191, 36, 0.1);
-    border: 1px solid rgba(251, 191, 36, 0.3);
-    border-radius: 6px;
-    padding: 0.75rem 1rem;
-    margin: 0.75rem 0;
-    font-size: 0.9rem;
-    color: #fbbf24;
-}
-[data-theme="light"] .warning { color: #92400e; background: rgba(251, 191, 36, 0.12); border-color: rgba(251, 191, 36, 0.4); }
-.warning ul {
-    margin: 0.5rem 0 0 1.25rem;
-    font-size: 0.85rem;
-}
-.warning li { margin-bottom: 0.25rem; }
-.external-addr {
-    font-size: 0.85rem;
-    color: var(--text-secondary);
-    margin-top: 0.5rem;
-}
-.external-addr.muted { opacity: 0.6; font-style: italic; }
-.nat-stat {
-    font-size: 0.85rem;
-    color: var(--text-secondary);
-    margin-top: 0.5rem;
-}
-.nat-fail { color: #f87171; font-weight: 500; }
-[data-theme="light"] .nat-fail { color: #dc2626; }
-.nat-advice {
-    background: rgba(248, 113, 113, 0.1);
-    border: 1px solid rgba(248, 113, 113, 0.3);
-    border-radius: 6px;
-    padding: 0.5rem 0.75rem;
-    margin-top: 0.5rem;
-    font-size: 0.85rem;
-    color: #f87171;
-}
-[data-theme="light"] .nat-advice { color: #b91c1c; background: rgba(248, 113, 113, 0.08); }
-.nat-recent { color: var(--text-muted); font-size: 0.8rem; }
-.nat-verdict {
-    display: inline-block;
-    font-size: 0.75rem;
-    font-weight: 600;
-    padding: 0.1rem 0.4rem;
-    border-radius: 4px;
-    margin-left: 0.3rem;
-    vertical-align: middle;
-}
-.nat-verdict-good { background: rgba(52, 211, 153, 0.15); color: #34d399; }
-.nat-verdict-bad { background: rgba(248, 113, 113, 0.15); color: #f87171; }
-.nat-verdict-warn { background: rgba(251, 191, 36, 0.15); color: #fbbf24; }
-[data-theme="light"] .nat-verdict-good { color: #059669; background: rgba(5, 150, 105, 0.1); }
-[data-theme="light"] .nat-verdict-bad { color: #dc2626; background: rgba(220, 38, 38, 0.1); }
-[data-theme="light"] .nat-verdict-warn { color: #d97706; background: rgba(217, 119, 6, 0.1); }
-.diagnostics {
-    background: rgba(251, 191, 36, 0.1);
-    border: 1px solid rgba(251, 191, 36, 0.3);
-    border-radius: 6px;
-    padding: 0.75rem 1rem;
-    margin-top: 0.75rem;
-}
-.diagnostics h3 {
-    color: #fbbf24;
-    font-size: 0.9rem;
-    margin-bottom: 0.4rem;
-}
-[data-theme="light"] .diagnostics h3 { color: #92400e; }
-[data-theme="light"] .diagnostics { color: #78350f; background: rgba(251, 191, 36, 0.12); }
-.diagnostics ul {
-    padding-left: 1.2rem;
-    margin: 0.4rem 0;
-    list-style: disc;
-}
-.diagnostics li {
-    color: var(--text-secondary);
-    margin-bottom: 0.35rem;
-    font-size: 0.85rem;
-}
-/* Stale-assets warning bar (#4289): shown by the JS only when the served
-   page's compile-time version differs from the live runtime version. */
-.version-banner {
-    background: rgba(248, 113, 113, 0.15);
-    border-bottom: 2px solid rgba(248, 113, 113, 0.5);
-    color: #f87171;
-    padding: 0.6rem 1.5rem;
-    font-size: 0.9rem;
-    font-weight: 500;
-    text-align: center;
-}
-[data-theme="light"] .version-banner {
-    color: #b91c1c;
-    background: rgba(220, 38, 38, 0.1);
-    border-bottom-color: rgba(220, 38, 38, 0.4);
-}
-.attempts {
-    color: var(--text-muted);
-    font-size: 0.8rem;
-    margin-top: 0.4rem;
-}
-.table-wrap { overflow-x: auto; }
-table {
-    width: 100%;
-    border-collapse: collapse;
-    font-size: 0.85rem;
-    font-family: var(--font-mono);
-}
-thead th {
-    text-align: left;
-    padding: 0.4rem 0.6rem;
-    border-bottom: 2px solid var(--border-color);
-    color: var(--text-secondary);
-    font-weight: 500;
-    font-size: 0.8rem;
-    text-transform: uppercase;
-    letter-spacing: 0.03em;
-}
-tbody td {
-    padding: 0.4rem 0.6rem;
-    border-bottom: 1px solid var(--border-color);
-}
-code {
-    background: var(--bg-tertiary);
-    color: var(--text-primary);
-    padding: 0.1rem 0.3rem;
-    border-radius: 3px;
-    font-size: 0.85em;
-    font-family: var(--font-mono);
-}
-.peer-row { cursor: pointer; transition: background 0.15s; }
-.peer-row:hover { background: var(--bg-tertiary); }
-.empty {
-    color: var(--text-muted);
-    font-size: 0.9rem;
-    font-style: italic;
-}
-.op-grid {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 0.75rem;
-}
-.op-cell {
-    text-align: center;
-    padding: 0.5rem;
-    background: var(--bg-secondary);
-    border-radius: 6px;
-    border: 1px solid var(--border-color);
-}
-.op-name {
-    font-size: 0.75rem;
-    color: var(--text-secondary);
-    font-weight: 500;
-    text-transform: uppercase;
-    margin-bottom: 0.25rem;
-    font-family: var(--font-mono);
-}
-.op-ok { color: #34d399; font-weight: 600; }
-.op-ok::before { content: "\2713 "; }
-.op-fail { color: #f87171; font-weight: 600; margin-left: 0.5rem; }
-.op-fail::before { content: "\2717 "; }
-[data-theme="light"] .op-ok { color: #059669; }
-[data-theme="light"] .op-fail { color: #dc2626; }
-.op-count { color: var(--text-primary); font-weight: 600; font-size: 1.1rem; }
-.op-received { color: var(--text-secondary); font-size: 0.7rem; margin-top: 0.15rem; }
-.app-list, .link-list {
-    list-style: none;
-    padding: 0;
-}
-.app-list li, .link-list li { margin-bottom: 0.4rem; }
-.app-list a, .link-list a {
-    color: var(--accent-light);
-    text-decoration: none;
-}
-.app-list a:hover, .link-list a:hover { text-decoration: underline; }
-[data-theme="light"] .app-list a, [data-theme="light"] .link-list a { color: var(--accent-primary); }
-.note {
-    color: var(--text-secondary);
-    font-size: 0.8rem;
-    margin-top: 0.15rem;
-}
-/* ── Health banner ── */
-.health-banner {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    padding: 0.6rem 0.9rem;
-    border-radius: 6px;
-    font-weight: 500;
-    font-size: 0.9rem;
-    margin-bottom: 0.75rem;
-}
-.health-icon { font-size: 1.1rem; flex-shrink: 0; }
-.health-good {
-    background: rgba(52, 211, 153, 0.12);
-    border: 1px solid rgba(52, 211, 153, 0.3);
-    color: #34d399;
-}
-[data-theme="light"] .health-good { color: #059669; background: rgba(52, 211, 153, 0.1); }
-.health-degraded {
-    background: rgba(251, 191, 36, 0.1);
-    border: 1px solid rgba(251, 191, 36, 0.3);
-    color: #fbbf24;
-}
-[data-theme="light"] .health-degraded { color: #92400e; }
-.health-connecting {
-    background: rgba(10, 186, 181, 0.08);
-    border: 1px solid rgba(10, 186, 181, 0.2);
-    color: var(--accent-light);
-}
-[data-theme="light"] .health-connecting { color: var(--accent-dark); }
-.health-trouble {
-    background: rgba(248, 113, 113, 0.1);
-    border: 1px solid rgba(248, 113, 113, 0.3);
-    color: #f87171;
-}
-[data-theme="light"] .health-trouble { color: #b91c1c; }
-/* ── Demoted diagnostics (when connected) ── */
-.diagnostics-muted {
-    margin-top: 0.5rem;
-    font-size: 0.8rem;
-    color: var(--text-muted);
-}
-.diagnostics-muted summary {
-    cursor: pointer;
-    font-family: var(--font-mono);
-}
-.diagnostics-muted summary:hover { color: var(--text-secondary); }
-.diagnostics-muted ul {
-    padding-left: 1.2rem;
-    margin: 0.4rem 0;
-    list-style: disc;
-    color: var(--text-secondary);
-    font-size: 0.8rem;
-}
-.diagnostics-muted li { margin-bottom: 0.25rem; }
-.muted-hint { opacity: 0.6; font-style: italic; }
-/* ── Ring stats metric row ── */
-.metrics-row {
-    display: flex;
-    gap: 0.5rem;
-    margin: 0.75rem 0;
-}
-.metric-tile {
-    flex: 1;
-    background: var(--bg-secondary);
-    border: 1px solid var(--border-color);
-    border-radius: 6px;
-    padding: 0.5rem;
-    text-align: center;
-}
-.metric-value {
-    display: block;
-    font-size: 1.3rem;
-    font-weight: 700;
-    font-family: var(--font-mono);
-    color: var(--accent-light);
-}
-[data-theme="light"] .metric-value { color: var(--accent-primary); }
-.metric-label {
-    display: block;
-    font-size: 0.65rem;
-    color: var(--text-secondary);
-    text-transform: uppercase;
-    font-family: var(--font-mono);
-    margin-top: 0.2rem;
-}
-/* ── Transfer stats ── */
-.transfer-stat {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0.5rem;
-    background: var(--bg-secondary);
-    border-radius: 6px;
-    border: 1px solid var(--border-color);
-}
-.transfer-label {
-    font-size: 0.75rem;
-    color: var(--text-secondary);
-    font-weight: 500;
-    text-transform: uppercase;
-    font-family: var(--font-mono);
-}
-.transfer-value {
-    font-size: 1.1rem;
-    font-weight: 600;
-    font-family: var(--font-mono);
-    color: var(--text-primary);
-}
-.transfer-detail {
-    font-size: 0.75rem;
-    font-weight: 400;
-    color: var(--text-secondary);
-    margin-left: 0.4rem;
-}
-.note a { color: var(--accent-light); }
-[data-theme="light"] .note a { color: var(--accent-primary); }
-.ext-link { color: var(--accent-light); }
-[data-theme="light"] .ext-link { color: var(--accent-primary); }
-p { margin-bottom: 0.5rem; }
-p:last-child { margin-bottom: 0; }
-.ring-wrap {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-bottom: 0.75rem;
-}
-.ring-svg { display: block; }
-.ring-svg circle:first-child { stroke: var(--text-muted) !important; }
-.ring-legend {
-    display: flex;
-    gap: 1rem;
-    margin-top: 0.4rem;
-    font-size: 0.75rem;
-    color: var(--text-secondary);
-}
-.ring-key {
-    display: flex;
-    align-items: center;
-    gap: 0.3rem;
-}
-.ring-dot {
-    width: 10px;
-    height: 10px;
-    border-radius: 50%;
-    display: inline-block;
-}
-.ring-dot-self { background: #43c178; }
-.ring-dot-peer { background: #66d9ff; }
-.ring-dot-gw { background: #ffb610; }
-.ring-dot-contract-normal { background: #43c178; opacity: 0.55; }
-.ring-dot-contract-flagged { background: #ff667a; }
-.theme-btn {
-    background: none;
-    border: 1px solid var(--border-color);
-    border-radius: 6px;
-    padding: 0.2rem 0.5rem;
-    cursor: pointer;
-    font-size: 1rem;
-    line-height: 1;
-    color: inherit;
-    transition: border-color 0.15s;
-}
-.theme-btn:hover { border-color: var(--text-secondary); }
-
-/* ── Mobile responsive ── */
-@media (max-width: 768px) {
-    main { margin: 0.75rem auto; padding: 0 0.5rem; }
-    header {
-        flex-wrap: wrap; gap: 0.4rem; padding: 0.5rem 0.75rem;
-    }
-    .header-left { flex-wrap: wrap; gap: 0.3rem; }
-    .header-right { margin-left: auto; }
-    .pub-key-label { display: none; }
-    .pub-key { max-width: 10ch; font-size: 0.65rem; }
-    .copy-btn { font-size: 0.65rem; padding: 0.15rem 0.3rem; }
-    .uptime { font-size: 0.7rem; }
-    .badge, .update-badge { font-size: 0.65rem; }
-
-    .card { padding: 0.75rem 1rem; border-radius: 8px; }
-
-    .metrics-row { flex-wrap: wrap; gap: 0.35rem; }
-    .metric-tile { flex: 1 1 30%; min-width: 5rem; padding: 0.4rem; }
-    .metric-value { font-size: 1.1rem; }
-    .metric-label { font-size: 0.6rem; }
-
-    .peer-table { font-size: 0.72rem; }
-    .peer-table th, .peer-table td { padding: 0.3rem 0.4rem; }
-    .peer-table th:nth-child(4), .peer-table td:nth-child(4),
-    .peer-table th:nth-child(5), .peer-table td:nth-child(5) { display: none; }
-
-    .transfer-stat { flex-direction: column; align-items: flex-start; gap: 0.2rem; }
-    .transfer-value { font-size: 0.95rem; }
-
-    .op-grid { grid-template-columns: repeat(2, 1fr); gap: 0.35rem; }
-    .op-card { padding: 0.5rem; }
-    .op-label { font-size: 0.6rem; }
-    .op-value { font-size: 0.9rem; }
-
-    .ring-wrap { margin-bottom: 0.5rem; }
-    .ring-svg { width: 160px; height: 160px; }
-
-    /* Governance: stack verdict + norms vertically, let table scroll */
-    .g-verdict-row { grid-template-columns: 1fr; }
-    .g-verdict { min-width: auto; }
-    .g-norms { grid-template-columns: repeat(3, 1fr); }
-    .app-list li { padding: 0.4rem 0; }
-}
-@media (max-width: 400px) {
-    .g-norms { grid-template-columns: repeat(2, 1fr); }
-    .header-title { font-size: 0.9rem; }
-    .header-scope { font-size: 0.65rem; padding: 0.1rem 0.35rem; }
-    .metric-tile { flex: 1 1 45%; }
-    .peer-table { font-size: 0.68rem; }
-    .op-grid { grid-template-columns: 1fr 1fr; }
-    .metrics-row { gap: 0.25rem; }
-}
-/* ── Ring SVG peer links ── */
-.ring-svg a.ring-peer-link { cursor: pointer; }
-.ring-svg a.ring-peer-link circle {
-    transition: r 0.12s ease, stroke-width 0.12s ease;
-}
-.ring-svg a.ring-peer-link:hover circle,
-.ring-svg a.ring-peer-link:focus circle {
-    r: 7;
-    stroke: var(--text-primary);
-    stroke-width: 1.5;
-}
-.ring-svg a.ring-peer-link:focus { outline: none; }
-
-/* "You" marker: subtle pulse */
-.ring-svg .you-dot {
-    animation: you-pulse 3s ease-in-out infinite;
-    transform-origin: center;
-}
-
-/* Peer table: fade-in rows */
-.peer-table tbody tr {
-    animation: peer-fade-in 0.35s ease-out both;
-}
-.peer-table tbody tr:nth-child(1) { animation-delay: 0.02s; }
-.peer-table tbody tr:nth-child(2) { animation-delay: 0.06s; }
-.peer-table tbody tr:nth-child(3) { animation-delay: 0.10s; }
-.peer-table tbody tr:nth-child(4) { animation-delay: 0.14s; }
-.peer-table tbody tr:nth-child(5) { animation-delay: 0.18s; }
-.peer-table tbody tr:nth-child(n+6) { animation-delay: 0.22s; }
-
-.copy-btn-inline {
-    font-family: var(--font-mono);
-    font-size: 0.65rem;
-    color: var(--text-muted);
-    background: none;
-    border: 1px solid var(--border-color);
-    border-radius: 3px;
-    padding: 0 0.35rem;
-    cursor: pointer;
-    margin-left: 0.4rem;
-    vertical-align: middle;
-}
-.copy-btn-inline:hover { color: var(--accent-light); border-color: var(--accent-light); }
-/* ── Sortable tables ── */
-table.sortable thead th {
-    cursor: pointer;
-    user-select: none;
-    position: relative;
-    padding-right: 1.1rem;
-}
-table.sortable thead th:hover { color: var(--text-primary); }
-table.sortable thead th::after {
-    content: "";
-    position: absolute;
-    right: 0.4rem;
-    top: 50%;
-    transform: translateY(-50%);
-    opacity: 0.35;
-    font-size: 0.7rem;
-}
-table.sortable thead th.sort-asc::after { content: "▲"; opacity: 1; color: var(--accent-light); }
-table.sortable thead th.sort-desc::after { content: "▼"; opacity: 1; color: var(--accent-light); }
-/* ── Copy-to-clipboard contract button ── */
-.copy-key {
-    background: none;
-    border: none;
-    padding: 0 0.25rem;
-    margin-left: 0.35rem;
-    cursor: copy;
-    font: inherit;
-    font-size: 0.85em;
-    color: var(--text-muted);
-    transition: color 0.15s;
-}
-.copy-key:hover { color: var(--accent-light); }
-.copy-key:focus { outline: 1px dashed var(--accent-light); outline-offset: 2px; }
-.copy-key.copied { color: #34d399; }
-[data-theme="light"] .copy-key.copied { color: #059669; }
-/* ── Update-available badge in header ── */
-.update-badge {
-    background: rgba(52, 211, 153, 0.15);
-    color: #34d399;
-    border: 1px solid rgba(52, 211, 153, 0.35);
-    padding: 0.15rem 0.5rem;
-    border-radius: 12px;
-    font-size: 0.75rem;
-    font-weight: 500;
-    font-family: var(--font-mono);
-    text-decoration: none;
-    margin-left: 0.25rem;
-}
-.update-badge:hover { background: rgba(52, 211, 153, 0.25); text-decoration: none; }
-[data-theme="light"] .update-badge { color: #059669; }
-.update-badge[hidden] { display: none; }
-/* ── Toast (clipboard feedback) ── */
-.toast-container {
-    position: fixed;
-    bottom: 1rem;
-    left: 50%;
-    transform: translateX(-50%);
-    z-index: 100;
-    display: flex;
-    flex-direction: column;
-    gap: 0.4rem;
-    pointer-events: none;
-}
-.toast {
-    background: var(--bg-panel);
-    border: 1px solid var(--border-color);
-    color: var(--text-primary);
-    padding: 0.5rem 0.9rem;
-    border-radius: 6px;
-    font-size: 0.85rem;
-    font-family: var(--font-mono);
-    backdrop-filter: blur(10px);
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.25);
-    animation: toast-in 0.18s ease-out;
-}
-.toast.toast-error { color: #f87171; border-color: rgba(248, 113, 113, 0.4); }
-@keyframes toast-in {
-    from { opacity: 0; transform: translateY(8px); }
-    to   { opacity: 1; transform: translateY(0); }
-}
-
-/* ── Governance card (Phase 4.5) ───────────────────────────────────── */
-.g-mode {
-    font-family: var(--font-mono);
-    font-size: 0.7rem;
-    padding: 0.15rem 0.5rem;
-    border-radius: 3px;
-    letter-spacing: 0.03em;
-    text-transform: uppercase;
-    margin-left: auto;
-}
-.g-mode-dry-run {
-    background: rgba(255, 182, 16, 0.15);
-    color: #ffb610;
-    border: 1px solid rgba(255, 182, 16, 0.35);
-}
-.g-mode-enforce {
-    background: rgba(255, 102, 122, 0.15);
-    color: #ff667a;
-    border: 1px solid rgba(255, 102, 122, 0.35);
-}
-.g-mode-off {
-    background: rgba(148, 163, 184, 0.15);
-    color: var(--text-secondary);
-    border: 1px solid rgba(148, 163, 184, 0.35);
-}
-
-.g-verdict-row {
-    display: grid;
-    grid-template-columns: minmax(280px, 0.85fr) 2fr;
-    gap: 0.75rem;
-    margin: 0.5rem 0 0.75rem;
-}
-.g-verdict {
-    background: var(--bg-tertiary);
-    border: 1px solid var(--border-color);
-    border-left: 3px solid #ff8a3d;
-    border-radius: 6px;
-    padding: 0.85rem 1rem 0.85rem 1.1rem;
-    display: grid;
-    grid-template-columns: auto 1fr;
-    grid-template-rows: auto auto;
-    column-gap: 1rem;
-    align-items: center;
-}
-.verdict-ok { border-left-color: #43c178; }
-.verdict-alert { border-left-color: #ff8a3d; }
-.g-verdict .verdict-num {
-    grid-row: 1 / 3;
-    font-family: var(--font-mono);
-    font-size: 3.2rem;
-    font-weight: 500;
-    line-height: 0.95;
-    color: #ff8a3d;
-    letter-spacing: -0.04em;
-}
-.verdict-ok .verdict-num { color: #43c178; }
-.g-verdict .verdict-headline {
-    font-size: 0.95rem;
-    font-weight: 500;
-    color: var(--text-primary);
-}
-.g-verdict .verdict-detail {
-    font-family: var(--font-mono);
-    font-size: 0.78rem;
-    color: var(--text-secondary);
-    margin-top: 0.15rem;
-}
-.g-verdict .sw {
-    display: inline-block;
-    width: 6px;
-    height: 6px;
-    border-radius: 50%;
-    margin-right: 0.3rem;
-    vertical-align: 0.07rem;
-}
-.sw-borderline { background: #ffb610; }
-.sw-wouldevict { background: #ff8a3d; }
-.sw-evicted    { background: #ff667a; }
-.sw-banned     { background: #d33682; }
-
-.g-norms {
-    display: grid;
-    grid-template-columns: repeat(5, 1fr);
-    gap: 0.5rem;
-}
-.g-norm {
-    background: var(--bg-tertiary);
-    border: 1px solid var(--border-color);
-    border-radius: 5px;
-    padding: 0.55rem 0.75rem;
-}
-.g-norm-label {
-    font-family: var(--font-mono);
-    font-size: 0.66rem;
-    text-transform: uppercase;
-    letter-spacing: 0.04em;
-    color: var(--text-muted);
-}
-.g-norm-value {
-    font-family: var(--font-mono);
-    font-size: 1rem;
-    font-weight: 500;
-    margin-top: 0.1rem;
-    color: var(--text-primary);
-}
-
-.g-badge {
-    display: inline-block;
-    font-family: var(--font-mono);
-    font-size: 0.7rem;
-    padding: 0.1rem 0.45rem;
-    border-radius: 3px;
-    text-transform: uppercase;
-    letter-spacing: 0.03em;
-    font-weight: 500;
-}
-.g-normal     { background: rgba(67,193,120,0.16);  color: #43c178; }
-.g-borderline { background: rgba(255,182,16,0.16);  color: #ffb610; }
-.g-wouldevict { background: rgba(255,138,61,0.18);  color: #ff8a3d; }
-.g-evicted    { background: rgba(255,102,122,0.18); color: #ff667a; }
-.g-banned     { background: rgba(211,54,130,0.18);  color: #d33682; }
-
-/* Compact governance-state pill rendered inline in the Subscribed
- * Contracts table's Gov column. Smaller than the `.g-badge` used in
- * the main Contract Governance table so it doesn't dominate the
- * row, while still being a single visual cue scanning down the
- * column. */
-.gov-pill {
-    display: inline-block;
-    padding: 0.1rem 0.45rem;
-    font-size: 0.72rem;
-    border-radius: 3px;
-    text-transform: uppercase;
-    letter-spacing: 0.04em;
-    font-weight: 500;
-    line-height: 1.2;
-}
-.gov-ok         { background: rgba(67,193,120,0.12);  color: #6b8a76; }
-.gov-borderline { background: rgba(255,182,16,0.18);  color: #ffb610; }
-.gov-wouldevict { background: rgba(255,138,61,0.20);  color: #ff8a3d; }
-.gov-evicted    { background: rgba(255,102,122,0.22); color: #ff667a; }
-.gov-banned     { background: rgba(211,54,130,0.22);  color: #d33682; }
-
-/* Denominator in the empty-state verdict ("3/30"). Smaller than the
- * numerator so the eye reads the count first and the threshold
- * second. */
-.verdict-num-denom {
-    font-size: 0.55em;
-    color: var(--text-muted, #888);
-    margin-left: 0.1em;
-    vertical-align: middle;
-}
-"##;
-
-/// Inline JavaScript for the dark/light mode toggle.
-/// Dark mode is the default (matching the global telemetry dashboard).
-/// A plain (non-module) script so it runs synchronously before first paint
-/// and so toggleTheme() is reachable from the button's onclick attribute.
-/// The page auto-refreshes every 5 s, so restoring the saved theme before
-/// render is essential to avoid a flash of the wrong theme on each refresh.
-const JS: &str = r##"
-(function() {
-    try {
-        if (localStorage.getItem('theme') === 'light') {
-            document.documentElement.setAttribute('data-theme', 'light');
-        }
-    } catch (e) { /* localStorage unavailable — default to dark */ }
-})();
-
-function toggleTheme() {
-    var isLight = document.documentElement.getAttribute('data-theme') === 'light';
-    var icon = document.getElementById('theme-icon');
-    if (isLight) {
-        document.documentElement.removeAttribute('data-theme');
-        if (icon) icon.textContent = '\u2600\uFE0F'; /* sun = click to switch to light */
-        try { localStorage.removeItem('theme'); } catch (e) {}
-    } else {
-        document.documentElement.setAttribute('data-theme', 'light');
-        if (icon) icon.textContent = '\uD83C\uDF19'; /* moon = click to switch to dark */
-        try { localStorage.setItem('theme', 'light'); } catch (e) {}
-    }
-}
-
-/* ── Toast notifications ── */
-function showToast(msg, opts) {
-    var container = document.getElementById('toast-container');
-    if (!container) {
-        container = document.createElement('div');
-        container.id = 'toast-container';
-        container.className = 'toast-container';
-        document.body.appendChild(container);
-    }
-    var t = document.createElement('div');
-    t.className = 'toast' + ((opts && opts.error) ? ' toast-error' : '');
-    t.textContent = msg;
-    container.appendChild(t);
-    setTimeout(function() {
-        t.style.transition = 'opacity 0.25s';
-        t.style.opacity = '0';
-        setTimeout(function() { if (t.parentNode) t.parentNode.removeChild(t); }, 260);
-    }, (opts && opts.duration) || 1600);
-}
-
-/* ── Copy contract key to clipboard ── */
-function copyToClipboard(text) {
-    if (navigator.clipboard && navigator.clipboard.writeText) {
-        return navigator.clipboard.writeText(text);
-    }
-    /* Fallback for older browsers / non-secure contexts */
-    return new Promise(function(resolve, reject) {
-        try {
-            var ta = document.createElement('textarea');
-            ta.value = text;
-            ta.style.position = 'fixed';
-            ta.style.opacity = '0';
-            document.body.appendChild(ta);
-            ta.select();
-            var ok = document.execCommand('copy');
-            document.body.removeChild(ta);
-            ok ? resolve() : reject(new Error('execCommand failed'));
-        } catch (e) { reject(e); }
-    });
-}
-
-/* ── Sortable tables ── */
-function compareCells(a, b, type) {
-    if (type === 'num') {
-        var na = parseFloat(a);
-        var nb = parseFloat(b);
-        var aBad = isNaN(na), bBad = isNaN(nb);
-        if (aBad && bBad) return 0;
-        if (aBad) return 1;   /* missing values sort to bottom */
-        if (bBad) return -1;
-        return na - nb;
-    }
-    return a.localeCompare(b, undefined, { numeric: true, sensitivity: 'base' });
-}
-
-function applySort(table, colIndex, dir) {
-    var tbody = table.querySelector('tbody');
-    if (!tbody) return;
-    var ths = table.querySelectorAll('thead th');
-    var th = ths[colIndex];
-    if (!th) return;
-    var type = th.getAttribute('data-sort-type') || 'text';
-    var rows = Array.prototype.slice.call(tbody.querySelectorAll('tr'));
-    rows.sort(function(r1, r2) {
-        var c1 = r1.children[colIndex];
-        var c2 = r2.children[colIndex];
-        var v1 = c1 ? (c1.getAttribute('data-sort') || c1.textContent) : '';
-        var v2 = c2 ? (c2.getAttribute('data-sort') || c2.textContent) : '';
-        var cmp = compareCells(v1, v2, type);
-        return dir === 'desc' ? -cmp : cmp;
-    });
-    rows.forEach(function(r) { tbody.appendChild(r); });
-    ths.forEach(function(h) { h.classList.remove('sort-asc', 'sort-desc'); });
-    th.classList.add(dir === 'desc' ? 'sort-desc' : 'sort-asc');
-}
-
-function sortKey(table) {
-    return 'sort:' + (table.getAttribute('data-table-id') || 'tbl');
-}
-
-function handleHeaderClick(th) {
-    var table = th.closest('table.sortable');
-    if (!table) return;
-    var ths = Array.prototype.slice.call(table.querySelectorAll('thead th'));
-    var idx = ths.indexOf(th);
-    if (idx < 0) return;
-    var current = th.classList.contains('sort-asc') ? 'asc'
-                : th.classList.contains('sort-desc') ? 'desc'
-                : null;
-    var dir = current === 'asc' ? 'desc' : 'asc';
-    applySort(table, idx, dir);
-    try { sessionStorage.setItem(sortKey(table), idx + ':' + dir); } catch (e) {}
-}
-
-function restoreSort() {
-    document.querySelectorAll('table.sortable').forEach(function(table) {
-        try {
-            var saved = sessionStorage.getItem(sortKey(table));
-            if (!saved) return;
-            var parts = saved.split(':');
-            var idx = parseInt(parts[0], 10);
-            var dir = parts[1] === 'desc' ? 'desc' : 'asc';
-            if (!isNaN(idx)) applySort(table, idx, dir);
-        } catch (e) {}
-    });
-}
-
-/* ── Update-available check (GitHub releases, cached 12h) ── */
-function compareSemver(a, b) {
-    var pa = String(a).replace(/^v/, '').split(/[.\-+]/);
-    var pb = String(b).replace(/^v/, '').split(/[.\-+]/);
-    var n = Math.max(pa.length, pb.length);
-    for (var i = 0; i < n; i++) {
-        var na = parseInt(pa[i], 10);
-        var nb = parseInt(pb[i], 10);
-        if (isNaN(na) && isNaN(nb)) {
-            var s = (pa[i] || '').localeCompare(pb[i] || '');
-            if (s !== 0) return s;
-            continue;
-        }
-        if (isNaN(na)) return -1;
-        if (isNaN(nb)) return 1;
-        if (na !== nb) return na - nb;
-    }
-    return 0;
-}
-
-function showUpdateBadge(latestTag) {
-    var el = document.getElementById('update-badge');
-    if (!el) return;
-    el.textContent = 'Update: v' + String(latestTag).replace(/^v/, '');
-    el.title = 'A newer Freenet release is available — click to view release notes';
-    el.hidden = false;
-}
-
-function checkForUpdate() {
-    var badge = document.getElementById('version-badge');
-    if (!badge) return;
-    var current = badge.getAttribute('data-version') || '';
-    if (!current || current === '?') return;
-    var TTL_MS = 12 * 60 * 60 * 1000;
-    var now = Date.now();
-    var cached = null;
-    try {
-        var raw = localStorage.getItem('freenet-update-check');
-        if (raw) cached = JSON.parse(raw);
-    } catch (e) {}
-    if (cached && cached.tag && cached.checkedAt && (now - cached.checkedAt) < TTL_MS) {
-        if (compareSemver(cached.tag, current) > 0) showUpdateBadge(cached.tag);
-        return;
-    }
-    fetch('https://api.github.com/repos/freenet/freenet-core/releases/latest', {
-        headers: { 'Accept': 'application/vnd.github+json' },
-    }).then(function(r) {
-        if (!r.ok) throw new Error('HTTP ' + r.status);
-        return r.json();
-    }).then(function(data) {
-        var tag = data && data.tag_name;
-        if (!tag) return;
-        try { localStorage.setItem('freenet-update-check', JSON.stringify({ tag: tag, checkedAt: now })); } catch (e) {}
-        if (compareSemver(tag, current) > 0) showUpdateBadge(tag);
-    }).catch(function(e) {
-        /* Network blocked / GitHub rate-limited — silently skip */
-        console.debug('Update check failed:', e);
-    });
-}
-
-/* A version string is "known" when it is non-empty and not the '?'
-   placeholder the homepage uses before a node snapshot exists.
-   Mirrors version_is_known() in home_page.rs — keep both in sync. */
-function versionIsKnown(v) {
-    return !!v && v !== '?';
-}
-
-/* Show the stale-assets banner iff both the asset version (baked into this
-   served page at compile time) and the live runtime version are known and
-   differ. Mirrors should_show_version_banner() in home_page.rs. The point
-   of comparing against a LIVE fetch (not the rendered data-version) is to
-   catch the #4289 case: the browser is holding a cached page emitted by an
-   old binary while a newer binary is now answering requests. */
-function checkVersionMismatch() {
-    var banner = document.getElementById('version-mismatch-banner');
-    if (!banner) return;
-    var assetVersion = banner.getAttribute('data-asset-version') || '';
-    if (!versionIsKnown(assetVersion)) return;
-    fetch('/v1/version', { headers: { 'Accept': 'application/json' } }).then(function(r) {
-        if (!r.ok) throw new Error('HTTP ' + r.status);
-        return r.json();
-    }).then(function(data) {
-        var runtimeVersion = data && data.version;
-        if (!versionIsKnown(runtimeVersion)) return;
-        if (runtimeVersion !== assetVersion) {
-            banner.textContent = 'Asset version ' + assetVersion + ' ≠ node version '
-                + runtimeVersion + ' — this page is stale, refresh to load the current version.';
-            banner.hidden = false;
-        } else {
-            /* Versions agree (e.g. after a refresh fixed the staleness). */
-            banner.hidden = true;
-        }
-    }).catch(function(e) {
-        /* Endpoint unreachable / node mid-startup — don't show a spurious banner. */
-        console.debug('Version check failed:', e);
-    });
-}
-
-/* Tab switching for per-operation-type charts */
-function switchTab(el) {
-    var tabId = el.getAttribute('data-tab');
-    /* Deactivate all tabs and panels in this group */
-    var group = el.closest('.tab-group');
-    if (!group) return;
-    group.querySelectorAll('.tab-label').forEach(function(t) { t.classList.remove('tab-active'); });
-    group.querySelectorAll('.tab-panel').forEach(function(p) { p.classList.remove('tab-panel-active'); });
-    /* Activate selected */
-    el.classList.add('tab-active');
-    var panel = group.querySelector('#panel-' + tabId);
-    if (panel) panel.classList.add('tab-panel-active');
-    /* Remember active tab for auto-refresh persistence */
-    try { sessionStorage.setItem('activeOpTab', tabId); } catch(e) {}
-}
-
-document.addEventListener('DOMContentLoaded', function() {
-    var icon = document.getElementById('theme-icon');
-    if (icon && document.documentElement.getAttribute('data-theme') === 'light') {
-        icon.textContent = '\uD83C\uDF19'; /* moon = click to switch to dark */
-    }
-
-    /* Restore active tab after page load / auto-refresh */
-    function restoreTab() {
-        try {
-            var saved = sessionStorage.getItem('activeOpTab');
-            if (saved) {
-                var tab = document.querySelector('.tab-label[data-tab="' + saved + '"]');
-                if (tab) switchTab(tab);
-            }
-        } catch(e) {}
-    }
-    restoreTab();
-    restoreSort();
-    checkForUpdate();
-    checkVersionMismatch();
-
-    /* Delegated click handler \u2014 survives <main> innerHTML swaps from auto-refresh,
-       so we don't need to re-bind after each refresh. */
-    document.addEventListener('click', function(ev) {
-        var copy = ev.target.closest && ev.target.closest('.copy-key');
-        if (copy) {
-            ev.preventDefault();
-            ev.stopPropagation();
-            var text = copy.getAttribute('data-copy') || copy.textContent.trim();
-            copyToClipboard(text).then(function() {
-                showToast('Contract key copied');
-                copy.classList.add('copied');
-                setTimeout(function() { copy.classList.remove('copied'); }, 900);
-            }).catch(function() {
-                showToast('Copy failed', { error: true });
-            });
-            return;
-        }
-        var th = ev.target.closest && ev.target.closest('table.sortable thead th');
-        if (th) { handleHeaderClick(th); return; }
-    });
-
-    /* Auto-refresh: fetch the page and swap dynamic content without a full reload.
-       Uses setTimeout chaining (not setInterval) so slow responses don't overlap. */
-    function scheduleRefresh() {
-        setTimeout(function() {
-            fetch(window.location.href).then(function(r) { return r.text(); }).then(function(html) {
-                var parser = new DOMParser();
-                var doc = parser.parseFromString(html, 'text/html');
-                var newMain = doc.querySelector('main');
-                var oldMain = document.querySelector('main');
-                if (newMain && oldMain) oldMain.innerHTML = newMain.innerHTML;
-                /* Update header elements (outside <main>) */
-                var newUp = doc.querySelector('.uptime');
-                var oldUp = document.querySelector('.uptime');
-                if (newUp && oldUp) oldUp.textContent = newUp.textContent;
-                var newBadge = doc.querySelector('#version-badge');
-                var oldBadge = document.getElementById('version-badge');
-                if (newBadge && oldBadge) {
-                    oldBadge.textContent = newBadge.textContent;
-                    var nv = newBadge.getAttribute('data-version');
-                    if (nv) oldBadge.setAttribute('data-version', nv);
-                }
-                var newIcon = doc.querySelector('link[rel="icon"]');
-                var oldIcon = document.querySelector('link[rel="icon"]');
-                if (newIcon && oldIcon) oldIcon.setAttribute('href', newIcon.getAttribute('href'));
-                /* Restore tab selection and table sort after content swap */
-                restoreTab();
-                restoreSort();
-                /* Re-check the live runtime version so the stale-assets banner
-                   appears (or clears) if the serving process changes while the
-                   page stays open. The banner's data-asset-version stays anchored
-                   to the originally-loaded page, which is the version we're
-                   comparing against. */
-                checkVersionMismatch();
-            }).catch(function(e) { console.warn('Dashboard refresh failed:', e); })
-              .finally(scheduleRefresh);
-        }, 5000);
-    }
-    scheduleRefresh();
-});
-"##;
-
-// ─── Peer detail page ────────────────────────────────────────────────────────
-
-fn peer_detail_html(address_str: &str) -> String {
-    let snap = network_status::get_snapshot();
-
-    let peer = snap.as_ref().and_then(|s| {
-        s.peers
-            .iter()
-            .find(|p| p.address.to_string() == address_str)
-    });
-
-    let Some(peer) = peer else {
-        return format!(
-            r##"<!DOCTYPE html>
-<html lang="en"><head>
-<meta charset="UTF-8"><meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>Peer Not Found — Freenet</title>
-<style>{CSS}{PEER_CSS}</style><script>{JS}</script>
-</head><body>
-<header>
-    <div class="header-left">
-        <a href="/" class="logo-link"><img src="https://freenet.org/freenet_logo.svg" alt="Freenet" class="logo"></a>
-        <a href="/" class="header-title">FREENET</a>
-        <span class="header-sep">/</span>
-        <span class="header-scope">Peer</span>
-    </div>
-    <div class="header-right">
-        <button class="theme-btn" id="theme-btn" onclick="toggleTheme()" title="Toggle dark/light mode">
-            <span id="theme-icon">☀️</span>
-        </button>
-    </div>
-</header>
-<main>
-    <div class="card"><h2>Peer Not Found</h2><p class="empty">No connected peer with address <code>{addr}</code>. The peer may have disconnected.</p>
-    <p style="margin-top:0.75rem"><a href="/" style="color:var(--accent-light);font-family:var(--font-mono);font-size:0.85rem">&larr; Back to dashboard</a></p></div>
-</main></body></html>"##,
-            CSS = CSS,
-            PEER_CSS = PEER_CSS,
-            JS = JS,
-            addr = html_escape(address_str),
-        );
-    };
-
-    let peer_type = if peer.is_gateway { "Gateway" } else { "Peer" };
-    let loc_str = peer
-        .location
-        .map(|l| format!("{:.6}", l))
-        .unwrap_or_else(|| "—".to_string());
-
-    // Try to get router data
-    let router_lock = network_status::get_router();
-    let router_guard = router_lock.as_ref().map(|r| r.read());
-
-    let (router_snapshot, peer_routing) = match (&router_guard, &peer.peer_key_location) {
-        (Some(router), Some(pkl)) => (Some(router.snapshot()), Some(router.peer_snapshot(pkl))),
-        (Some(router), None) => (Some(router.snapshot()), None),
-        _ => (None, None),
-    };
-
-    // Build info card
-    let addr_enc = html_escape(&peer.address.to_string());
-    let info_card = format!(
-        r#"<div class="card">
-            <h2>Peer Info</h2>
-            <div><strong>{ptype}</strong> <code>{addr}</code><button type="button" class="copy-btn-inline" data-addr="{addr_enc}" onclick="copyToClipboard(this.getAttribute('data-addr')).then(function(){{showToast('Address copied')}})" title="Copy address">⎘</button></div>
-            <div class="info-grid">
-                <div class="info-label">Location</div><div class="info-value">{loc}</div>
-                <div class="info-label">Connected</div><div class="info-value">{connected}</div>
-                <div class="info-label">Sent</div><div class="info-value">{sent}</div>
-                <div class="info-label">Received</div><div class="info-value">{recv}</div>
-            </div>
-        </div>"#,
-        ptype = peer_type,
-        addr = addr_enc,
-        addr_enc = addr_enc,
-        loc = loc_str,
-        connected = format_duration(peer.connected_secs),
-        sent = format_bytes(peer.bytes_sent),
-        recv = format_bytes(peer.bytes_received),
-    );
-
-    // Build routing model status card
-    let model_card = if let Some(ref rs) = router_snapshot {
-        let total_events = rs.failure_events + rs.success_events;
-        let peer_failure_events = peer_routing
-            .as_ref()
-            .and_then(|pr| pr.failure_adjustment.map(|(_, c)| c))
-            .unwrap_or(0);
-        let peer_response_events = peer_routing
-            .as_ref()
-            .and_then(|pr| pr.response_time_adjustment.map(|(_, c)| c))
-            .unwrap_or(0);
-        let peer_transfer_events = peer_routing
-            .as_ref()
-            .and_then(|pr| pr.transfer_rate_adjustment.map(|(_, c)| c))
-            .unwrap_or(0);
-        format!(
-            r#"<div class="card">
-                <h2>Routing Model</h2>
-                <div class="info-grid">
-                    <div class="info-label">Prediction active</div><div class="info-value">{active}</div>
-                    <div class="info-label">Global events</div><div class="info-value">{total}</div>
-                    <div class="info-label">This peer: failure</div><div class="info-value">{pf} events</div>
-                    <div class="info-label">This peer: response time</div><div class="info-value">{pr} events</div>
-                    <div class="info-label">This peer: transfer rate</div><div class="info-value">{pt} events</div>
-                </div>
-                <h3 style="margin-top: 1em;">Renegade ML Predictor</h3>
-                <p class="empty" style="font-size: 0.8em; margin-top: 0.25em;"><a href="https://github.com/sanity/renegade" target="_blank" rel="noopener noreferrer" class="ext-link">Renegade</a> is a zero-configuration k-nearest-neighbours model (it auto-selects K and learns which features matter). It predicts per-peer, per-contract outcomes from four features (peer, contract location, distance, time); the router blends its prediction with the distance-based estimate (a weighted average, Renegade's share growing with data up to half) to catch patterns distance alone misses, such as a peer that drops requests for specific contracts.</p>
-                <div class="info-grid">
-                    <div class="info-label">Failure observations</div><div class="info-value">{rf}</div>
-                    <div class="info-label">Response time observations</div><div class="info-value">{rr}</div>
-                    <div class="info-label">Transfer speed observations</div><div class="info-value">{rt}</div>
-                    <div class="info-label">Known peers</div><div class="info-value">{rp}</div>
-                    <div class="info-label">Predictions evaluated</div><div class="info-value">{n_eval}</div>
-                    <div class="info-label">Brier score (overall)</div><div class="info-value">{brier}</div>
-                    <div class="info-label">Brier score (recent)</div><div class="info-value">{recent_brier}</div>
-                </div>
-                <p class="empty" style="font-size: 0.8em; margin-top: 0.5em;">Brier score: 0 = perfect · &lt;0.05 excellent · &lt;0.1 good · &lt;0.15 fair · 0.25 = random coin flip</p>
-            </div>"#,
-            active = if rs.prediction_active { "Yes" } else { "No" },
-            total = total_events,
-            pf = peer_failure_events,
-            pr = peer_response_events,
-            pt = peer_transfer_events,
-            rf = rs.renegade_failure_events,
-            rr = rs.renegade_response_time_events,
-            rt = rs.renegade_transfer_speed_events,
-            rp = rs.renegade_known_peers,
-            brier = rs
-                .renegade_brier_score
-                .map(|b| format!("{:.4}", b))
-                .unwrap_or_else(|| "—".to_string()),
-            recent_brier = rs
-                .renegade_recent_brier_score
-                .map(|b| format!("{:.4}", b))
-                .unwrap_or_else(|| "—".to_string()),
-            n_eval = rs.renegade_predictions_evaluated,
-        )
-    } else {
-        r#"<div class="card"><h2>Routing Model</h2><p class="empty">Router data not available</p></div>"#.to_string()
-    };
-
-    // Build SVG charts with per-operation-type tabs
-    let charts = if let Some(ref rs) = router_snapshot {
-        let fail_adj = peer_routing
-            .as_ref()
-            .and_then(|pr| pr.failure_adjustment.map(|(m, _)| m));
-        let rt_adj = peer_routing
-            .as_ref()
-            .and_then(|pr| pr.response_time_adjustment.map(|(m, _)| m));
-        let xfer_adj = peer_routing
-            .as_ref()
-            .and_then(|pr| pr.transfer_rate_adjustment.map(|(m, _)| m));
-        let ploc = peer.location;
-
-        // Build tab content for each operation type
-        let tab_names = ["All", "GET", "PUT", "UPDATE", "SUBSCRIBE"];
-        let mut tab_labels = String::new();
-        let mut tab_panels = String::new();
-
-        for (i, &tab_name) in tab_names.iter().enumerate() {
-            let tab_id = tab_name.to_lowercase().replace(' ', "-");
-
-            // Get curves + raw scatter points for this tab
-            #[allow(clippy::type_complexity)]
-            let (
-                f_curve,
-                f_range,
-                f_points,
-                rt_curve,
-                rt_range,
-                rt_points,
-                xfer_curve,
-                xfer_range,
-                xfer_points,
-                event_count,
-            ): (
-                &[(f64, f64)],
-                (f64, f64),
-                &[(f64, f64)],
-                &[(f64, f64)],
-                (f64, f64),
-                &[(f64, f64)],
-                &[(f64, f64)],
-                (f64, f64),
-                &[(f64, f64)],
-                usize,
-            ) = if tab_name == "All" {
-                (
-                    rs.failure_curve.as_slice(),
-                    rs.failure_data_range,
-                    rs.failure_points.as_slice(),
-                    rs.response_time_curve.as_slice(),
-                    rs.response_time_data_range,
-                    rs.response_time_points.as_slice(),
-                    rs.transfer_rate_curve.as_slice(),
-                    rs.transfer_rate_data_range,
-                    rs.transfer_rate_points.as_slice(),
-                    rs.failure_events,
-                )
-            } else if let Some(c) = rs.per_op_curves.get(tab_name) {
-                (
-                    c.failure_curve.as_slice(),
-                    c.failure_data_range,
-                    c.failure_points.as_slice(),
-                    c.response_time_curve.as_slice(),
-                    c.response_time_data_range,
-                    c.response_time_points.as_slice(),
-                    c.transfer_rate_curve.as_slice(),
-                    c.transfer_rate_data_range,
-                    c.transfer_rate_points.as_slice(),
-                    c.failure_events,
-                )
-            } else {
-                (
-                    &[][..],
-                    (0.0, 0.0),
-                    &[][..],
-                    &[][..],
-                    (0.0, 0.0),
-                    &[][..],
-                    &[][..],
-                    (0.0, 0.0),
-                    &[][..],
-                    0,
-                )
-            };
-
-            // Tab label with event count badge
-            let count_badge = if event_count > 0 {
-                format!(r#" <span class="tab-count">{event_count}</span>"#)
-            } else {
-                String::new()
-            };
-            let dim_class = if event_count == 0 && tab_name != "All" {
-                " tab-dim"
-            } else {
-                ""
-            };
-            let active_class = if i == 0 { " tab-active" } else { "" };
-            write!(
-                tab_labels,
-                r#"<span class="tab-label{dim}{active}" data-tab="{id}" onclick="switchTab(this)">{name}{badge}</span>"#,
-                id = tab_id, dim = dim_class, active = active_class, name = tab_name, badge = count_badge,
-            ).ok();
-
-            // Tab panel content
-            let mut panel_content = String::new();
-            if event_count == 0 && tab_name != "All" {
-                write!(
-                    panel_content,
-                    r#"<div class="empty-chart">No {name} operations have routed through this peer yet. The chart will populate as the network sends or relays {name}s through this connection.</div>"#,
-                    name = tab_name,
-                )
-                .ok();
-            } else {
-                // Always render all three prediction-component slots so the
-                // user can see every dimension the router models. Each slot
-                // either contains the rendered curve or a per-metric
-                // "awaiting data" placeholder — empty slots are NOT hidden,
-                // because hiding them silently rotted unobserved when the
-                // migration stopped feeding the response-time
-                // and transfer-rate estimators (the `Failure Probability`-only
-                // dashboard regression that surfaced this code path).
-                // Failure probabilities are tiny, so a fixed 0.0–1.0 axis
-                // squashes the curve flat against the bottom. Zoom the top of
-                // the axis to 2x the curve's value at the right edge of the
-                // plot so the line is legible. See failure_chart_y_max.
-                let fail_y_max =
-                    failure_chart_y_max(f_curve, if tab_name == "All" { fail_adj } else { None })
-                        .to_string();
-                panel_content.push_str(&build_estimator_chart_or_placeholder(
-                    "Failure Probability",
-                    f_curve,
-                    f_points,
-                    f_range,
-                    if tab_name == "All" { fail_adj } else { None },
-                    ploc,
-                    "0.0",
-                    &fail_y_max,
-                    "No success/failure observations have routed through this peer yet.",
-                ));
-                panel_content.push_str(&build_estimator_chart_or_placeholder(
-                    "Response Time (s)",
-                    rt_curve,
-                    rt_points,
-                    rt_range,
-                    if tab_name == "All" { rt_adj } else { None },
-                    ploc,
-                    "0",
-                    "auto",
-                    "No timed responses have been observed from this peer yet.",
-                ));
-                panel_content.push_str(&build_estimator_chart_or_placeholder(
-                    "Transfer Rate (B/s)",
-                    xfer_curve,
-                    xfer_points,
-                    xfer_range,
-                    if tab_name == "All" { xfer_adj } else { None },
-                    ploc,
-                    // Transfer rate is a positive B/s value: floor at 0, auto-scale
-                    // the top. (Was "auto"/"0", which clamped the max to 0 and gave a
-                    // degenerate inverted range that also hid the scatter overlay.)
-                    "0",
-                    "auto",
-                    "No payload transfers have been observed from this peer yet.",
-                ));
-            }
-
-            let panel_active = if i == 0 { " tab-panel-active" } else { "" };
-            write!(
-                tab_panels,
-                r#"<div class="tab-panel{active}" id="panel-{id}">{content}</div>"#,
-                active = panel_active,
-                id = tab_id,
-                content = panel_content,
-            )
-            .ok();
-        }
-
-        format!(
-            r#"<div class="card">
-                <h2>Outcomes vs Distance</h2>
-                <p style="font-size:0.8em;color:var(--text-muted);">
-                    Actual observed outcomes (dots) against ring distance to the contract, with the
-                    isotonic fit overlaid (the All tab is the aggregate the router uses; per-op tabs
-                    just break it down and are not consulted separately). "Peer-adjusted" adds this
-                    peer's running EWMA correction to that fit. How tightly the dots hug a monotonic
-                    curve shows how well distance alone predicts the outcome. A separate Renegade
-                    model is blended into the final estimate; its accuracy is in the Prediction
-                    Accuracy panel below.
-                </p>
-                <p class="chart-legend">
-                    <span class="chart-key"><span class="chart-dot chart-dot-actual"></span> Actual outcomes</span>
-                    <span class="chart-key"><span class="chart-dot chart-dot-global"></span> Isotonic fit</span>
-                    <span class="chart-key"><span class="chart-dot chart-dot-peer"></span> Peer-adjusted</span>
-                    <span class="chart-key"><span class="chart-dot chart-dot-loc"></span> Peer location</span>
-                    <span class="chart-key"><span class="chart-dot chart-dot-ext"></span> Extrapolated</span>
-                </p>
-                <div class="tab-group">
-                    <div class="tab-bar">{tab_labels}</div>
-                    {tab_panels}
-                </div>
-            </div>"#,
-            tab_labels = tab_labels,
-            tab_panels = tab_panels,
-        )
-    } else {
-        String::new()
-    };
-
-    // Build the renegade prediction-accuracy panel (failure + timing models)
-    let renegade_chart = if let Some(ref rs) = router_snapshot {
-        build_renegade_accuracy_panel(
-            &rs.renegade_accuracy_pairs,
-            rs.renegade_brier_score,
-            &rs.renegade_response_time_pairs,
-            &rs.renegade_transfer_speed_pairs,
-        )
-    } else {
-        String::new()
-    };
-
-    // Build prediction summary card
-    let prediction_card = if let Some(ref pr) = peer_routing {
-        if let Some(ref pred) = pr.prediction_at_own_location {
-            format!(
-                r#"<div class="card">
-                    <h2>Prediction at Peer Location</h2>
-                    <div class="info-grid">
-                        <div class="info-label">Failure probability</div><div class="info-value">{fp}</div>
-                        <div class="info-label">Response time</div><div class="info-value">{rt}</div>
-                        <div class="info-label">Expected total time</div><div class="info-value">{ett}</div>
-                        <div class="info-label">Transfer speed</div><div class="info-value">{ts}</div>
-                    </div>
-                </div>"#,
-                fp = fmt_prediction_prob(pred.failure_probability),
-                rt = fmt_prediction_time(pred.time_to_response_start),
-                ett = fmt_prediction_time(pred.expected_total_time),
-                ts = fmt_prediction_speed(pred.transfer_speed_bps),
-            )
-        } else {
-            r#"<div class="card"><h2>Prediction</h2><p class="empty">Not enough routing data yet to predict this peer's behavior. The card fills in as operations are routed through it.</p></div>"#.to_string()
-        }
-    } else {
-        String::new()
-    };
-
-    let snap_ref = snap.as_ref();
-    let version = snap_ref.map(|s| s.version.as_str()).unwrap_or("?");
-
-    format!(
-        r##"<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Peer {addr} — Freenet</title>
-    <style>{CSS}{PEER_CSS}</style>
-    <script>{JS}</script>
-</head>
-<body>
-    <header>
-        <div class="header-left">
-            <a href="/" class="logo-link"><img src="https://freenet.org/freenet_logo.svg" alt="Freenet" class="logo"></a>
-            <a href="/" class="header-title">FREENET</a>
-            <span class="header-sep">/</span>
-            <span class="header-scope">Peer</span>
-            <span class="header-addr">{addr}</span>
-            <span class="badge">v{version}</span>
-        </div>
-        <div class="header-right">
-            <button class="theme-btn" id="theme-btn" onclick="toggleTheme()" title="Toggle dark/light mode">
-                <span id="theme-icon">☀️</span>
-            </button>
-        </div>
-    </header>
-    <main>
-        {info_card}
-        {model_card}
-        {charts}
-        {renegade_chart}
-        {prediction_card}
-    </main>
-</body>
-</html>"##,
-        addr = html_escape(&peer.address.to_string()),
-        CSS = CSS,
-        PEER_CSS = PEER_CSS,
-        JS = JS,
-        version = html_escape(version),
-        info_card = info_card,
-        model_card = model_card,
-        charts = charts,
-        renegade_chart = renegade_chart,
-        prediction_card = prediction_card,
-    )
-}
-
-/// Choose the top of the failure-probability chart's y-axis.
-///
-/// Failure probabilities for a healthy peer are tiny (often well under 0.1),
-/// so a fixed 0.0–1.0 axis squashes the fitted curve flat against the bottom
-/// and its shape is unreadable. Instead, scale the top of the axis to twice the
-/// curve's value at the right edge of the plot (the largest distance shown,
-/// 0.5), so the line occupies roughly the lower half of the chart. The PAV
-/// failure estimator is monotonically increasing in distance, so the right edge
-/// is the curve's maximum and 2x leaves headroom above the whole line.
-///
-/// Returns 1.0 (the original full-range axis) when there is no failure signal at
-/// the right edge, avoiding a degenerate zero-height axis. The result is capped
-/// at 1.0 since a probability can never exceed 1.0.
-fn failure_chart_y_max(curve_points: &[(f64, f64)], peer_adjustment: Option<f64>) -> f64 {
-    // y-value at the largest sampled distance (the right edge of the chart).
-    let right_edge = curve_points
-        .iter()
-        .max_by(|a, b| a.0.total_cmp(&b.0))
-        .map(|&(_, y)| y)
-        .unwrap_or(0.0);
-    // Keep the peer-adjusted line on-screen too: only an upward (positive)
-    // adjustment can push its right edge above the global curve's.
-    let right_edge = right_edge + peer_adjustment.unwrap_or(0.0).max(0.0);
-    if right_edge <= 1e-9 {
-        return 1.0;
-    }
-    (2.0 * right_edge).min(1.0)
-}
-
-/// Render the named estimator chart, or — when no data has been observed
-/// yet — a titled placeholder. The placeholder keeps the slot visible so
-/// users can always see every component of a routing prediction even when
-/// some estimators have not yet received feedback. Hiding empty charts
-/// masked the data-collection regression in the migration
-/// that fed only `failure_estimator` and left `response_start_time` and
-/// `transfer_rate` permanently empty.
-#[allow(clippy::too_many_arguments)]
-fn build_estimator_chart_or_placeholder(
-    title: &str,
-    curve_points: &[(f64, f64)],
-    scatter_points: &[(f64, f64)],
-    data_range: (f64, f64),
-    peer_adjustment: Option<f64>,
-    peer_location: Option<f64>,
-    y_min_hint: &str,
-    y_max_hint: &str,
-    empty_message: &str,
-) -> String {
-    if curve_points.is_empty() {
-        return format!(
-            r#"<div class="chart-section"><h3>{title}</h3><div class="empty-chart">{msg}</div></div>"#,
-            title = title,
-            msg = empty_message,
-        );
-    }
-    build_estimator_chart(
-        title,
-        curve_points,
-        scatter_points,
-        data_range,
-        peer_adjustment,
-        peer_location,
-        y_min_hint,
-        y_max_hint,
-    )
-}
-
-/// Build an SVG chart showing a PAV regression curve with optional per-peer adjustment.
-///
-/// `data_range` is `(data_x_min, data_x_max)` -- the x-range of actual regression data.
-/// Points outside this range are extrapolated by the PAV crate and drawn as dashed lines.
-#[allow(clippy::too_many_arguments)]
-fn build_estimator_chart(
-    title: &str,
-    curve_points: &[(f64, f64)],
-    scatter_points: &[(f64, f64)],
-    data_range: (f64, f64),
-    peer_adjustment: Option<f64>,
-    peer_location: Option<f64>,
-    y_min_hint: &str,
-    y_max_hint: &str,
-) -> String {
-    if curve_points.is_empty() {
-        return format!(
-            r#"<div class="chart-section"><h3>{title}</h3><div class="empty-chart">No data yet. Populates as operations route through this peer.</div></div>"#,
-            title = title,
-        );
-    }
-
-    let w: f64 = 560.0;
-    // Bottom padding leaves room for both the distance tick numbers and the
-    // "Distance" axis title below them; plot height stays 160px.
-    let h: f64 = 210.0;
-    let pad_l: f64 = 50.0;
-    let pad_r: f64 = 10.0;
-    let pad_t: f64 = 10.0;
-    let pad_b: f64 = 40.0;
-    let plot_w = w - pad_l - pad_r;
-    let plot_h = h - pad_t - pad_b;
-
-    // Determine Y range: use fixed bounds if provided, otherwise auto-scale from data
-    let fixed_y_min = y_min_hint.parse::<f64>().ok();
-    let fixed_y_max = y_max_hint.parse::<f64>().ok();
-
-    let mut y_min;
-    let mut y_max;
-
-    if let (Some(lo), Some(hi)) = (fixed_y_min, fixed_y_max) {
-        y_min = lo;
-        y_max = hi;
-    } else {
-        let y_vals: Vec<f64> = curve_points.iter().map(|(_, y)| *y).collect();
-        y_min = y_vals.iter().cloned().fold(f64::INFINITY, f64::min);
-        y_max = y_vals.iter().cloned().fold(f64::NEG_INFINITY, f64::max);
-
-        // Include the raw scatter so observed outliers aren't clipped.
-        for (_, y) in scatter_points {
-            if y.is_finite() {
-                y_min = y_min.min(*y);
-                y_max = y_max.max(*y);
-            }
-        }
-
-        // Include peer-adjusted values in range if present
-        if let Some(adj) = peer_adjustment {
-            for (_, y) in curve_points {
-                let adjusted = y + adj;
-                y_min = y_min.min(adjusted);
-                y_max = y_max.max(adjusted);
-            }
-        }
-
-        // Override individual bounds if a fixed hint was given
-        if let Some(lo) = fixed_y_min {
-            y_min = lo;
-        }
-        if let Some(hi) = fixed_y_max {
-            y_max = hi;
-        }
-
-        // Add 10% padding and avoid zero-range (only for auto-scaled bounds)
-        let range = y_max - y_min;
-        if range < 1e-10 {
-            y_min -= 0.5;
-            y_max += 0.5;
-        } else {
-            if fixed_y_min.is_none() {
-                y_min -= range * 0.1;
-            }
-            if fixed_y_max.is_none() {
-                y_max += range * 0.1;
-            }
-        }
-    }
-    let y_range = y_max - y_min;
-
-    // X is always distance [0.0, 0.5]
-    let x_min: f64 = 0.0;
-    let x_max: f64 = 0.5;
-    let x_range = x_max - x_min;
-
-    let to_svg_x = |x: f64| -> f64 { pad_l + ((x - x_min) / x_range) * plot_w };
-    let to_svg_y = |y: f64| -> f64 { pad_t + plot_h - ((y - y_min) / y_range) * plot_h };
-
-    let mut svg = format!(
-        r#"<div class="chart-section"><h3>{title}</h3>
-        <svg viewBox="0 0 {w} {h}" width="{w}" height="{h}" class="chart-svg">"#,
-        title = title,
-        w = w as u32,
-        h = h as u32,
-    );
-
-    // Axes
-    write!(
-        svg,
-        r#"<line x1="{lx}" y1="{ty}" x2="{lx}" y2="{by}" stroke="var(--text-muted)" stroke-width="1"/>"#,
-        lx = pad_l,
-        ty = pad_t,
-        by = pad_t + plot_h,
-    )
-    .ok();
-    write!(
-        svg,
-        r#"<line x1="{lx}" y1="{by}" x2="{rx}" y2="{by}" stroke="var(--text-muted)" stroke-width="1"/>"#,
-        lx = pad_l,
-        by = pad_t + plot_h,
-        rx = pad_l + plot_w,
-    )
-    .ok();
-
-    // X-axis labels
-    for &x_tick in &[0.0, 0.1, 0.2, 0.3, 0.4, 0.5] {
-        let sx = to_svg_x(x_tick);
-        write!(
-            svg,
-            r#"<text x="{sx:.0}" y="{y}" text-anchor="middle" class="axis-label">{v:.1}</text>"#,
-            sx = sx,
-            y = pad_t + plot_h + 18.0,
-            v = x_tick,
-        )
-        .ok();
-    }
-
-    // X-axis title: the x-axis is always ring distance (peer ↔ contract).
-    write!(
-        svg,
-        r#"<text x="{x:.0}" y="{y:.0}" text-anchor="middle" class="axis-label">Distance</text>"#,
-        x = pad_l + plot_w / 2.0,
-        y = h - 6.0,
-    )
-    .ok();
-
-    // Y-axis labels (3 ticks).
-    //
-    // Pick decimal places from the tick step so adjacent ticks stay
-    // distinguishable. The failure chart now zooms to a very small range
-    // (probabilities are tiny), where a fixed 2-decimal format would collapse
-    // every tick to "0.00".
-    let step = y_range / 2.0;
-    let decimals: usize = if step <= 0.0 {
-        3
-    } else if step >= 10.0 {
-        0
-    } else if step >= 1.0 {
-        1
-    } else {
-        // step in (0, 1): enough places for ~2 significant figures of the step.
-        ((-step.log10()).ceil() as usize).saturating_add(1).min(9)
-    };
-    for i in 0..=2 {
-        let frac = i as f64 / 2.0;
-        let y_val = y_min + frac * y_range;
-        let sy = to_svg_y(y_val);
-        let label = format!("{y_val:.decimals$}");
-        write!(
-            svg,
-            r#"<text x="{x}" y="{sy:.0}" text-anchor="end" class="axis-label">{label}</text>"#,
-            x = pad_l - 4.0,
-            sy = sy,
-            label = label,
-        )
-        .ok();
-    }
-
-    // Raw observed outcomes (drawn first, under the isotonic fit). Each dot is one
-    // actual event at its (distance, outcome); the spread shows how isotonic the
-    // relationship really is.
-    //
-    // The failure chart zooms its y-axis to the tiny fitted probabilities (see
-    // failure_chart_y_max), which would push the binary failure outcomes (y = 1.0)
-    // off the top of the plot. Rather than dropping off-scale points — which would
-    // hide every failure precisely on the healthy, low-probability peers the zoom
-    // is meant to illuminate — clamp them to the nearest edge so they remain
-    // visible as a row of dots at the boundary. (Auto-scaled charts always size
-    // their range to include the scatter, so the clamp is a no-op there.)
-    for &(x, y) in scatter_points {
-        if !(x.is_finite() && y.is_finite()) {
-            continue;
-        }
-        if !(x_min..=x_max).contains(&x) {
-            continue;
-        }
-        write!(
-            svg,
-            r#"<circle cx="{cx:.1}" cy="{cy:.1}" r="1.8" fill="var(--text-muted)" opacity="0.35"/>"#,
-            cx = to_svg_x(x),
-            cy = to_svg_y(y.clamp(y_min, y_max)),
-        )
-        .ok();
-    }
-
-    // Helper: draw a curve with solid line in data range and dashed outside.
-    // `points` are (x, y) pairs; `adj` is an optional y-offset for peer adjustment.
-    let draw_curve = |svg: &mut String, points: &[(f64, f64)], adj: f64, color: &str| {
-        if points.len() < 2 {
-            return;
-        }
-        let (data_lo, data_hi) = data_range;
-
-        // Split points into segments: extrapolated-left, data, extrapolated-right
-        let mut left_ext = Vec::new();
-        let mut data_seg = Vec::new();
-        let mut right_ext = Vec::new();
-
-        for &(x, y) in points {
-            let y = y + adj;
-            if x < data_lo - 0.001 {
-                left_ext.push((x, y));
-            } else if x > data_hi + 0.001 {
-                right_ext.push((x, y));
-            } else {
-                data_seg.push((x, y));
-            }
-        }
-
-        // Draw left extrapolation (dashed) -- include first data point for continuity
-        if !left_ext.is_empty() {
-            if let Some(&first_data) = data_seg.first() {
-                left_ext.push(first_data);
-            }
-            let mut path = String::new();
-            for (i, (x, y)) in left_ext.iter().enumerate() {
-                let sx = to_svg_x(*x);
-                let sy = to_svg_y(*y);
-                if i == 0 {
-                    write!(path, "M{sx:.1},{sy:.1}").ok();
-                } else {
-                    write!(path, " L{sx:.1},{sy:.1}").ok();
-                }
-            }
-            write!(
-                svg,
-                r#"<path d="{path}" fill="none" stroke="{color}" stroke-width="1.5" stroke-dasharray="4,3" opacity="0.5"/>"#,
-                path = path, color = color,
-            ).ok();
-        }
-
-        // Draw data range (solid)
-        if data_seg.len() >= 2 {
-            let mut path = String::new();
-            for (i, (x, y)) in data_seg.iter().enumerate() {
-                let sx = to_svg_x(*x);
-                let sy = to_svg_y(*y);
-                if i == 0 {
-                    write!(path, "M{sx:.1},{sy:.1}").ok();
-                } else {
-                    write!(path, " L{sx:.1},{sy:.1}").ok();
-                }
-            }
-            write!(
-                svg,
-                r#"<path d="{path}" fill="none" stroke="{color}" stroke-width="2" opacity="0.8"/>"#,
-                path = path,
-                color = color,
-            )
-            .ok();
-        } else if data_seg.len() == 1 {
-            // Single data point -- draw as a dot
-            let (x, y) = data_seg[0];
-            write!(
-                svg,
-                r#"<circle cx="{cx:.1}" cy="{cy:.1}" r="3" fill="{color}" opacity="0.8"/>"#,
-                cx = to_svg_x(x),
-                cy = to_svg_y(y),
-                color = color,
-            )
-            .ok();
-        }
-
-        // Draw right extrapolation (dashed) -- include last data point for continuity
-        if !right_ext.is_empty() {
-            if let Some(&last_data) = data_seg.last() {
-                right_ext.insert(0, last_data);
-            }
-            let mut path = String::new();
-            for (i, (x, y)) in right_ext.iter().enumerate() {
-                let sx = to_svg_x(*x);
-                let sy = to_svg_y(*y);
-                if i == 0 {
-                    write!(path, "M{sx:.1},{sy:.1}").ok();
-                } else {
-                    write!(path, " L{sx:.1},{sy:.1}").ok();
-                }
-            }
-            write!(
-                svg,
-                r#"<path d="{path}" fill="none" stroke="{color}" stroke-width="1.5" stroke-dasharray="4,3" opacity="0.5"/>"#,
-                path = path, color = color,
-            ).ok();
-        }
-    };
-
-    // Global curve (teal, the brand accent)
-    draw_curve(&mut svg, curve_points, 0.0, "var(--accent-primary)");
-
-    // Peer-adjusted curve (violet — deliberately off the teal/green family so it
-    // is not confused with the teal global curve)
-    if let Some(adj) = peer_adjustment {
-        draw_curve(&mut svg, curve_points, adj, "#8b5cf6");
-    }
-
-    // Peer location marker (vertical dashed line)
-    if let Some(loc) = peer_location {
-        // Distance from peer to itself is 0, but contracts near the peer have small distances.
-        // Mark the peer's ring location on the x-axis as distance=0 (leftmost).
-        let _ = loc; // The peer itself is at distance 0 from contracts at its own location
-        let sx = to_svg_x(0.0);
-        write!(
-            svg,
-            "<line x1=\"{sx:.1}\" y1=\"{ty}\" x2=\"{sx:.1}\" y2=\"{by}\" stroke=\"#fbbf24\" stroke-width=\"1.5\" stroke-dasharray=\"4,3\" opacity=\"0.7\"/>",
-            sx = sx,
-            ty = pad_t,
-            by = pad_t + plot_h,
-        )
-        .ok();
-    }
-
-    svg.push_str("</svg></div>");
-    svg
-}
-
-/// Sentinel values (f64::MAX / 2.0 ~ 9e307) indicate insufficient transfer data.
-/// Cap at ~31 years in seconds -- anything above is clearly not a real prediction.
-const REASONABLE_TIME_LIMIT: f64 = 1.0e9;
-
-fn fmt_prediction_time(v: f64) -> String {
-    if v.is_finite() && (0.0..REASONABLE_TIME_LIMIT).contains(&v) {
-        format!("{v:.3}s")
-    } else {
-        "N/A".to_string()
-    }
-}
-
-fn fmt_prediction_speed(v: f64) -> String {
-    if v.is_finite() && v > 0.0 {
-        format!("{v:.0} B/s")
-    } else {
-        "N/A".to_string()
-    }
-}
-
-fn fmt_prediction_prob(v: f64) -> String {
-    if v.is_finite() && (0.0..=1.0).contains(&v) {
-        format!("{v:.4}")
-    } else {
-        "N/A".to_string()
-    }
-}
-
-/// Which kind of regression model a scatter chart is rendering, controlling the
-/// axis unit formatting (durations vs throughput).
-#[derive(Clone, Copy)]
-enum RegKind {
-    Time,
-    Speed,
-}
-
-/// Build the renegade prediction-accuracy panel: one reliability diagram for the
-/// binary failure model plus predicted-vs-actual scatters for the two regression
-/// models (response time, transfer speed). Returns an empty string when no model
-/// has scored any predictions yet, so a fresh node shows nothing rather than an
-/// empty card.
-fn build_renegade_accuracy_panel(
-    failure_pairs: &[(f64, f64)],
-    brier: Option<f64>,
-    response_time_pairs: &[(f64, f64)],
-    transfer_speed_pairs: &[(f64, f64)],
-) -> String {
-    if failure_pairs.is_empty() && response_time_pairs.is_empty() && transfer_speed_pairs.is_empty()
-    {
-        return String::new();
-    }
-
-    let failure = build_reliability_chart(failure_pairs, brier);
-    let response = build_regression_chart("Response time", RegKind::Time, response_time_pairs);
-    let transfer = build_regression_chart("Transfer speed", RegKind::Speed, transfer_speed_pairs);
-
-    format!(
-        r#"<div class="card">
-        <h2>Prediction Accuracy</h2>
-        <p style="font-size:0.8em;color:var(--text-muted);">
-            How well the Renegade predictor's recent predictions matched reality (this scores
-            the Renegade k-NN layer, not the distance-only fit in Outcomes vs Distance above).
-            On the dashed diagonal predictions are perfect: for failure, predicted probability
-            equals the observed failure rate (calibration); for the timing models, predicted
-            equals actual.
-        </p>
-        <div style="display:flex;flex-wrap:wrap;gap:1rem;justify-content:flex-start;">
-            {failure}
-            {response}
-            {transfer}
-        </div>
-    </div>"#,
-        failure = failure,
-        response = response,
-        transfer = transfer,
-    )
-}
-
-/// Reliability (calibration) diagram for the binary failure model.
-/// X = predicted failure probability, Y = observed failure rate within each
-/// predicted-probability bin. Dots on the diagonal mean the model is calibrated;
-/// above the line it under-predicts failure, below it over-predicts.
-fn build_reliability_chart(pairs: &[(f64, f64)], brier: Option<f64>) -> String {
-    use std::fmt::Write;
-
-    let valid: Vec<(f64, f64)> = pairs
-        .iter()
-        .copied()
-        .filter(|(p, a)| p.is_finite() && a.is_finite())
-        .map(|(p, a)| (p.clamp(0.0, 1.0), a))
-        .collect();
-
-    if valid.is_empty() {
-        return mini_chart_placeholder("Failure (calibration)", "predicted prob vs observed rate");
-    }
-
-    let n = valid.len();
-    let n_bins = 10usize;
-    let mut bin_pred_sum = vec![0.0f64; n_bins];
-    let mut bin_fail = vec![0usize; n_bins];
-    let mut bin_total = vec![0usize; n_bins];
-    for (p, a) in &valid {
-        let bin = ((p * n_bins as f64) as usize).min(n_bins - 1);
-        bin_pred_sum[bin] += p;
-        bin_total[bin] += 1;
-        if *a >= 0.5 {
-            bin_fail[bin] += 1;
-        }
-    }
-
-    let (w, h) = (260.0f64, 220.0f64);
-    let (pad_l, pad_r, pad_t, pad_b) = (38.0f64, 12.0f64, 30.0f64, 30.0f64);
-    let plot_w = w - pad_l - pad_r;
-    let plot_h = h - pad_t - pad_b;
-    let to_x = |v: f64| pad_l + v.clamp(0.0, 1.0) * plot_w;
-    let to_y = |v: f64| pad_t + (1.0 - v.clamp(0.0, 1.0)) * plot_h;
-
-    let mut svg = format!(
-        r#"<svg viewBox="0 0 {w} {h}" width="{w}" height="{h}" class="accuracy-chart">"#,
-        w = w as u32,
-        h = h as u32,
-    );
-
-    write!(
-        svg,
-        r#"<text x="{x}" y="14" font-size="10" font-weight="600" fill="var(--text-secondary)">Failure (calibration)</text>"#,
-        x = pad_l,
-    )
-    .ok();
-    let headline = match brier {
-        Some(b) if b.is_finite() => format!("Brier {b:.3} · n={n}"),
-        _ => format!("n={n}"),
-    };
-    write!(
-        svg,
-        r#"<text x="{x}" y="26" font-size="9" fill="var(--text-muted)">{headline}</text>"#,
-        x = pad_l,
-    )
-    .ok();
-
-    write!(
-        svg,
-        r#"<rect x="{lx}" y="{ty}" width="{pw}" height="{ph}" fill="var(--bg-secondary)" rx="2"/>"#,
-        lx = pad_l,
-        ty = pad_t,
-        pw = plot_w,
-        ph = plot_h,
-    )
-    .ok();
-
-    // perfect-calibration diagonal
-    write!(
-        svg,
-        r#"<line x1="{x1:.1}" y1="{y1:.1}" x2="{x2:.1}" y2="{y2:.1}" stroke="var(--text-muted)" stroke-width="1" stroke-dasharray="4"/>"#,
-        x1 = to_x(0.0),
-        y1 = to_y(0.0),
-        x2 = to_x(1.0),
-        y2 = to_y(1.0),
-    )
-    .ok();
-
-    // axis ticks at 0 / 0.5 / 1 on both axes
-    for &v in &[0.0_f64, 0.5, 1.0] {
-        write!(
-            svg,
-            r#"<text x="{x:.1}" y="{y:.1}" text-anchor="middle" font-size="8" fill="var(--text-muted)">{v}</text>"#,
-            x = to_x(v),
-            y = pad_t + plot_h + 12.0,
-        )
-        .ok();
-        write!(
-            svg,
-            r#"<text x="{x:.1}" y="{y:.1}" text-anchor="end" font-size="8" fill="var(--text-muted)">{v}</text>"#,
-            x = pad_l - 4.0,
-            y = to_y(v) + 3.0,
-        )
-        .ok();
-    }
-
-    // calibration curve through non-empty bins (in predicted-probability order)
-    let mut pts: Vec<(f64, f64, usize)> = Vec::new();
-    for b in 0..n_bins {
-        if bin_total[b] == 0 {
-            continue;
-        }
-        let mean_pred = bin_pred_sum[b] / bin_total[b] as f64;
-        let obs_rate = bin_fail[b] as f64 / bin_total[b] as f64;
-        pts.push((mean_pred, obs_rate, bin_total[b]));
-    }
-    if pts.len() >= 2 {
-        let path = pts
-            .iter()
-            .map(|(px, py, _)| format!("{:.1},{:.1}", to_x(*px), to_y(*py)))
-            .collect::<Vec<_>>()
-            .join(" ");
-        write!(
-            svg,
-            r#"<polyline points="{path}" fill="none" stroke="var(--accent-primary, #58a6ff)" stroke-width="1.5" opacity="0.8"/>"#,
-        )
-        .ok();
-    }
-    let max_bin = bin_total.iter().copied().max().unwrap_or(1).max(1);
-    for (px, py, count) in &pts {
-        let r = 2.5 + 3.5 * (*count as f64 / max_bin as f64).sqrt();
-        write!(
-            svg,
-            r#"<circle cx="{cx:.1}" cy="{cy:.1}" r="{r:.1}" fill="var(--accent-primary, #58a6ff)" opacity="0.85"/>"#,
-            cx = to_x(*px),
-            cy = to_y(*py),
-        )
-        .ok();
-    }
-
-    write!(
-        svg,
-        r#"<text x="{x:.1}" y="{y:.1}" text-anchor="middle" font-size="8" fill="var(--text-muted)">predicted fail prob</text>"#,
-        x = pad_l + plot_w / 2.0,
-        y = h - 1.0,
-    )
-    .ok();
-
-    svg.push_str("</svg>");
-    svg
-}
-
-/// Predicted-vs-actual scatter for a regression model (response time, transfer
-/// speed) on log-log axes, since both targets span orders of magnitude. Points
-/// on the dashed diagonal mean predicted == actual; the headline is the median
-/// absolute percentage error over the retained window.
-fn build_regression_chart(label: &str, kind: RegKind, pairs: &[(f64, f64)]) -> String {
-    use std::fmt::Write;
-
-    // Log axes require strictly-positive, finite values.
-    let valid: Vec<(f64, f64)> = pairs
-        .iter()
-        .copied()
-        .filter(|(p, a)| p.is_finite() && a.is_finite() && *p > 0.0 && *a > 0.0)
-        .collect();
-
-    if valid.len() < 2 {
-        return mini_chart_placeholder(label, "predicted vs actual");
-    }
-
-    let n = valid.len();
-
-    // Median absolute percentage error (robust to the heavy tails of latency /
-    // throughput) as the single headline number.
-    let mut apes: Vec<f64> = valid.iter().map(|(p, a)| ((p - a) / a).abs()).collect();
-    apes.sort_by(|x, y| x.partial_cmp(y).unwrap_or(std::cmp::Ordering::Equal));
-    let mid = apes.len() / 2;
-    let mdape = if apes.len() % 2 == 0 {
-        (apes[mid - 1] + apes[mid]) / 2.0
-    } else {
-        apes[mid]
-    };
-
-    // Shared log range across predicted and actual so the diagonal is 45°.
-    let mut lo = f64::INFINITY;
-    let mut hi = f64::NEG_INFINITY;
-    for (p, a) in &valid {
-        lo = lo.min(p.min(*a));
-        hi = hi.max(p.max(*a));
-    }
-    let mut log_lo = lo.log10();
-    let mut log_hi = hi.log10();
-    if (log_hi - log_lo) < 0.5 {
-        // Pad a near-flat range so points don't all sit on one edge.
-        let center = (log_hi + log_lo) / 2.0;
-        log_lo = center - 0.5;
-        log_hi = center + 0.5;
-    } else {
-        let pad = (log_hi - log_lo) * 0.08;
-        log_lo -= pad;
-        log_hi += pad;
-    }
-    let span = (log_hi - log_lo).max(1e-9);
-
-    let (w, h) = (260.0f64, 220.0f64);
-    let (pad_l, pad_r, pad_t, pad_b) = (38.0f64, 12.0f64, 30.0f64, 30.0f64);
-    let plot_w = w - pad_l - pad_r;
-    let plot_h = h - pad_t - pad_b;
-    let to_x = |v: f64| pad_l + ((v.log10() - log_lo) / span) * plot_w;
-    let to_y = |v: f64| pad_t + (1.0 - (v.log10() - log_lo) / span) * plot_h;
-
-    let mut svg = format!(
-        r#"<svg viewBox="0 0 {w} {h}" width="{w}" height="{h}" class="accuracy-chart">"#,
-        w = w as u32,
-        h = h as u32,
-    );
-    write!(
-        svg,
-        r#"<text x="{x}" y="14" font-size="10" font-weight="600" fill="var(--text-secondary)">{label}</text>"#,
-        x = pad_l,
-    )
-    .ok();
-    write!(
-        svg,
-        r#"<text x="{x}" y="26" font-size="9" fill="var(--text-muted)">median err {pct:.0}% · n={n}</text>"#,
-        x = pad_l,
-        pct = mdape * 100.0,
-    )
-    .ok();
-
-    write!(
-        svg,
-        r#"<rect x="{lx}" y="{ty}" width="{pw}" height="{ph}" fill="var(--bg-secondary)" rx="2"/>"#,
-        lx = pad_l,
-        ty = pad_t,
-        pw = plot_w,
-        ph = plot_h,
-    )
-    .ok();
-
-    // Power-of-ten gridlines + axis labels. When all points fall within a single
-    // decade (the common steady-state case) there is no power-of-ten boundary in
-    // range, so fall back to labelling the axis endpoints rather than rendering an
-    // unlabelled axis.
-    let first_decade = log_lo.ceil() as i32;
-    let last_decade = log_hi.floor() as i32;
-    let tick_vals: Vec<f64> = if first_decade <= last_decade {
-        (first_decade..=last_decade)
-            .map(|d| 10f64.powi(d))
-            .collect()
-    } else {
-        vec![10f64.powf(log_lo), 10f64.powf(log_hi)]
-    };
-    for val in tick_vals {
-        let gx = to_x(val);
-        let gy = to_y(val);
-        write!(
-            svg,
-            r#"<line x1="{gx:.1}" y1="{ty:.1}" x2="{gx:.1}" y2="{by:.1}" stroke="var(--text-muted)" stroke-width="0.3" stroke-dasharray="3"/>"#,
-            ty = pad_t,
-            by = pad_t + plot_h,
-        )
-        .ok();
-        write!(
-            svg,
-            r#"<text x="{gx:.1}" y="{y:.1}" text-anchor="middle" font-size="8" fill="var(--text-muted)">{lbl}</text>"#,
-            y = pad_t + plot_h + 12.0,
-            lbl = fmt_reg_axis(kind, val),
-        )
-        .ok();
-        write!(
-            svg,
-            r#"<text x="{x:.1}" y="{gy:.1}" text-anchor="end" font-size="8" fill="var(--text-muted)">{lbl}</text>"#,
-            x = pad_l - 4.0,
-            lbl = fmt_reg_axis(kind, val),
-        )
-        .ok();
-    }
-
-    // Perfect diagonal (predicted == actual).
-    write!(
-        svg,
-        r#"<line x1="{x1:.1}" y1="{y1:.1}" x2="{x2:.1}" y2="{y2:.1}" stroke="var(--text-muted)" stroke-width="1" stroke-dasharray="4"/>"#,
-        x1 = to_x(10f64.powf(log_lo)),
-        y1 = to_y(10f64.powf(log_lo)),
-        x2 = to_x(10f64.powf(log_hi)),
-        y2 = to_y(10f64.powf(log_hi)),
-    )
-    .ok();
-
-    for (p, a) in &valid {
-        write!(
-            svg,
-            r#"<circle cx="{cx:.1}" cy="{cy:.1}" r="2.2" fill="var(--accent-primary, #58a6ff)" opacity="0.45"/>"#,
-            cx = to_x(*p),
-            cy = to_y(*a),
-        )
-        .ok();
-    }
-
-    write!(
-        svg,
-        r#"<text x="{x:.1}" y="{y:.1}" text-anchor="middle" font-size="8" fill="var(--text-muted)">predicted (x) vs actual (y)</text>"#,
-        x = pad_l + plot_w / 2.0,
-        y = h - 1.0,
-    )
-    .ok();
-
-    svg.push_str("</svg>");
-    svg
-}
-
-/// Compact axis label for a regression value: durations as s/ms/µs, throughput
-/// as B/KB/MB/GB per second.
-fn fmt_reg_axis(kind: RegKind, v: f64) -> String {
-    match kind {
-        RegKind::Time => {
-            if v >= 1.0 {
-                format!("{v:.0}s")
-            } else if v >= 0.001 {
-                format!("{:.0}ms", v * 1000.0)
-            } else {
-                format!("{:.0}µs", v * 1_000_000.0)
-            }
-        }
-        RegKind::Speed => {
-            if v >= 1e9 {
-                format!("{:.0}GB/s", v / 1e9)
-            } else if v >= 1e6 {
-                format!("{:.0}MB/s", v / 1e6)
-            } else if v >= 1e3 {
-                format!("{:.0}KB/s", v / 1e3)
-            } else {
-                format!("{v:.0}B/s")
-            }
-        }
-    }
-}
-
-/// A small placeholder chart shown while a model has too little data to plot.
-fn mini_chart_placeholder(label: &str, sub: &str) -> String {
-    let (w, h) = (260.0f64, 220.0f64);
-    format!(
-        r#"<svg viewBox="0 0 {w} {h}" width="{w}" height="{h}" class="accuracy-chart">
-        <text x="38" y="14" font-size="10" font-weight="600" fill="var(--text-secondary)">{label}</text>
-        <text x="{cx}" y="{cy}" text-anchor="middle" font-size="10" fill="var(--text-muted)">collecting data…</text>
-        <text x="{cx}" y="{cy2}" text-anchor="middle" font-size="8" fill="var(--text-muted)">{sub}</text>
-    </svg>"#,
-        w = w as u32,
-        h = h as u32,
-        cx = w / 2.0,
-        cy = h / 2.0,
-        cy2 = h / 2.0 + 14.0,
-    )
-}
-
-const PEER_CSS: &str = r##"
-.logo-link, .header-title { text-decoration: none; }
-.logo-link:hover, a.header-title:hover { opacity: 0.8; }
-a.header-title {
-    background: linear-gradient(135deg, var(--accent-light) 0%, var(--accent-primary) 50%, var(--accent-dark) 100%);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-    font-family: var(--font-mono);
-    font-weight: 600;
-    font-size: 1.1rem;
-    letter-spacing: -0.02em;
-}
-.header-sep {
-    color: var(--text-muted);
-    font-size: 1rem;
-    margin: 0 0.15rem;
-}
-.header-addr {
-    font-family: var(--font-mono);
-    font-size: 0.8rem;
-    color: var(--text-secondary);
-    background: var(--bg-tertiary);
-    padding: 0.1rem 0.45rem;
-    border-radius: 4px;
-}
-.info-grid {
-    display: grid;
-    grid-template-columns: auto 1fr;
-    gap: 0.4rem 1.25rem;
-    font-size: 0.85rem;
-}
-.info-label {
-    color: var(--text-secondary);
-    font-family: var(--font-mono);
-    font-weight: 500;
-    text-transform: uppercase;
-    font-size: 0.75rem;
-    align-self: center;
-}
-.info-value { font-family: var(--font-mono); }
-.chart-section { margin: 1rem 0; }
-.chart-section h3 {
-    font-size: 0.8rem;
-    color: var(--text-secondary);
-    font-family: var(--font-mono);
-    font-weight: 500;
-    text-transform: uppercase;
-    margin-bottom: 0.5rem;
-}
-.chart-svg {
-    display: block;
-    width: 100%;
-    max-width: 600px;
-}
-/* Small-multiple accuracy charts: fixed width so the three sit side by side
-   in the flex row and wrap on narrow viewports (overridden global width:100%). */
-.accuracy-chart {
-    display: block;
-    width: 260px;
-    max-width: 100%;
-    height: auto;
-    flex: 0 0 auto;
-}
-.chart-svg .axis-label {
-    font-family: var(--font-mono);
-    font-size: 10px;
-    fill: var(--text-muted);
-}
-.chart-legend {
-    display: flex;
-    gap: 1rem;
-    margin-bottom: 0.75rem;
-    font-size: 0.75rem;
-    color: var(--text-secondary);
-}
-.chart-key {
-    display: flex;
-    align-items: center;
-    gap: 0.3rem;
-}
-.chart-dot {
-    width: 10px;
-    height: 10px;
-    border-radius: 50%;
-    display: inline-block;
-}
-.chart-dot-actual { background: var(--text-muted); opacity: 0.55; }
-.chart-dot-global { background: var(--accent-primary); }
-.chart-dot-peer { background: #8b5cf6; }
-.chart-dot-loc { background: #fbbf24; }
-.chart-dot-ext {
-    background: transparent;
-    border: 2px dashed var(--accent-primary);
-    width: 8px;
-    height: 8px;
-}
-/* Tabs */
-.tab-group { margin-top: 0.5rem; }
-.tab-bar {
-    display: flex;
-    gap: 0;
-    border-bottom: 1px solid var(--border-color);
-    margin-bottom: 0.75rem;
-    overflow-x: auto;
-}
-.tab-label {
-    font-family: var(--font-mono);
-    font-size: 0.75rem;
-    font-weight: 500;
-    text-transform: uppercase;
-    letter-spacing: 0.03em;
-    color: var(--text-secondary);
-    padding: 0.4rem 0.75rem;
-    cursor: pointer;
-    border-bottom: 2px solid transparent;
-    transition: color 0.15s, border-color 0.15s;
-    white-space: nowrap;
-    display: inline-flex;
-    align-items: center;
-    gap: 0.35rem;
-}
-.tab-label:hover { color: var(--text-primary); }
-.tab-label.tab-dim { color: var(--text-muted); }
-.tab-label.tab-dim:hover { color: var(--text-secondary); }
-.tab-label.tab-active {
-    color: var(--accent-primary);
-    border-bottom-color: var(--accent-primary);
-}
-.tab-count {
-    font-size: 0.65rem;
-    background: var(--bg-tertiary);
-    color: var(--text-muted);
-    padding: 0.05rem 0.35rem;
-    border-radius: 8px;
-    min-width: 1.2em;
-    text-align: center;
-}
-.tab-label.tab-active .tab-count {
-    background: rgba(10, 186, 181, 0.15);
-    color: var(--accent-primary);
-}
-.tab-panel { display: none; }
-.tab-panel.tab-panel-active { display: block; }
-.empty-chart {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 80px;
-    background: var(--bg-secondary);
-    border: 1px dashed var(--border-color);
-    border-radius: 6px;
-    color: var(--text-muted);
-    font-size: 0.8rem;
-    font-family: var(--font-mono);
-}
-"##;
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    // Explicit imports from submodules (not re-exported from root).
+    use super::assets::JS;
+    use super::cards::{
+        build_ban_list_card, build_contracts_card, build_governance_card, build_ops_card,
+        build_peers_card, build_ring_svg, build_status_card, build_transfer_card, format_bytes,
+        format_last_evaluated,
+    };
+    use super::estimator::{
+        RegKind, build_estimator_chart, build_estimator_chart_or_placeholder,
+        build_regression_chart, build_reliability_chart, build_renegade_accuracy_panel,
+        failure_chart_y_max, fmt_prediction_prob, fmt_prediction_speed, fmt_prediction_time,
+    };
+    use super::favicon::build_favicon_data_uri;
+    use super::peer_detail::peer_detail_html;
     use crate::node::network_status::{
         FailureSnapshot, HealthLevel, NatStatsSnapshot, NetworkStatusSnapshot, OpStatsSnapshot,
         RingStatsSnapshot,
@@ -5150,13 +1132,8 @@ mod tests {
     /// builder.
     #[test]
     fn peer_detail_panel_calls_estimator_helper_for_all_three_components() {
-        let src = include_str!("home_page.rs");
-        // Truncate at the test marker so the assertion below doesn't
-        // match against this very test's source.
-        let cutoff = src
-            .find("#[cfg(test)]")
-            .expect("home_page.rs must have a #[cfg(test)] section");
-        let prod = &src[..cutoff];
+        let src = include_str!("home_page/peer_detail.rs");
+        let prod = src;
         for title in [
             "Failure Probability",
             "Response Time (s)",
@@ -5322,13 +1299,9 @@ mod tests {
     #[test]
     fn peer_detail_links_renegade_to_repo() {
         // The "Renegade" label on the Routing Model card links to the project repo.
-        let src = include_str!("home_page.rs");
-        let cutoff = src
-            .find("#[cfg(test)]")
-            .expect("home_page.rs must have a #[cfg(test)] section");
-        let prod = &src[..cutoff];
+        let src = include_str!("home_page/peer_detail.rs");
         assert!(
-            prod.contains(r#"href="https://github.com/sanity/renegade""#),
+            src.contains(r#"href="https://github.com/sanity/renegade""#),
             "the Renegade label must link to https://github.com/sanity/renegade"
         );
     }
